Max-Min Scaling Function

This function allows you to scale vectors or an entire data frame using the max-min scaling method A numeric vector is always returned.

maxmin(x)

Arguments

x

Pass a vector or the required columns of a data frame through this argument.

Examples

library(magrittr)
library(dplyr)

rand <- rnorm(100, mean = 0, sd = 1)
data.frame(original = rand, transformed = maxmin(rand))
#>        original transformed
#> 1   -0.30763807   0.4113177
#> 2   -0.22314707   0.4305358
#> 3    0.19804653   0.5263396
#> 4   -0.98010859   0.2583591
#> 5    0.32368832   0.5549177
#> 6    1.47881212   0.8176595
#> 7   -1.07602808   0.2365415
#> 8   -1.53696096   0.1316988
#> 9    0.46765696   0.5876645
#> 10   0.92583126   0.6918798
#> 11  -0.35966083   0.3994847
#> 12  -0.12916321   0.4519132
#> 13  -0.87659557   0.2819039
#> 14  -1.16102760   0.2172077
#> 15  -1.07340586   0.2371379
#> 16   0.95335775   0.6981409
#> 17   1.78851505   0.8881039
#> 18   0.48302828   0.5911608
#> 19  -0.16219012   0.4444010
#> 20  -1.16559415   0.2161690
#> 21  -0.54821581   0.3565965
#> 22  -1.31462711   0.1822703
#> 23  -0.61504213   0.3413963
#> 24  -1.01337957   0.2507914
#> 25   0.53473768   0.6029225
#> 26   0.49786606   0.5945358
#> 27  -0.94803555   0.2656544
#> 28   0.76787058   0.6559504
#> 29  -0.02232209   0.4762150
#> 30  -0.21435106   0.4325366
#> 31   2.28045773   1.0000000
#> 32   0.16185810   0.5181082
#> 33   0.18357442   0.5230478
#> 34  -0.07753935   0.4636554
#> 35  -0.99982728   0.2538739
#> 36  -0.35105792   0.4014415
#> 37   0.47782816   0.5899780
#> 38  -0.33514803   0.4050604
#> 39   0.19757945   0.5262333
#> 40  -1.14975359   0.2197721
#> 41  -0.55031616   0.3561187
#> 42  -1.61977288   0.1128626
#> 43   0.96757478   0.7013747
#> 44  -0.27872831   0.4178935
#> 45  -1.42873571   0.1563154
#> 46   1.43992052   0.8088133
#> 47   0.10927697   0.5061482
#> 48  -0.27341323   0.4191024
#> 49   1.41142096   0.8023309
#> 50   1.31688552   0.7808281
#> 51   0.19060550   0.5246470
#> 52  -0.33727537   0.4045765
#> 53  -1.64905160   0.1062029
#> 54   0.16057330   0.5178160
#> 55  -0.12436564   0.4530044
#> 56   1.42955144   0.8064548
#> 57  -1.63097385   0.1103148
#> 58   0.18468296   0.5232999
#> 59   0.78832743   0.6606035
#> 60  -0.11944366   0.4541240
#> 61   1.92753415   0.9197248
#> 62  -0.38780767   0.3930825
#> 63  -0.80021960   0.2992763
#> 64   1.47426536   0.8166253
#> 65  -2.11596430   0.0000000
#> 66   0.81569917   0.6668294
#> 67   0.80278817   0.6638927
#> 68   0.36892315   0.5652068
#> 69   0.31680550   0.5533522
#> 70   1.17599100   0.7487805
#> 71  -1.00185648   0.2534124
#> 72   0.71549123   0.6440363
#> 73  -0.81899010   0.2950068
#> 74  -0.06310865   0.4669378
#> 75   1.62760020   0.8515025
#> 76  -0.11084419   0.4560800
#> 77   1.08559787   0.7282199
#> 78  -0.27063339   0.4197347
#> 79   1.39132890   0.7977608
#> 80   0.64956966   0.6290420
#> 81  -0.88639133   0.2796758
#> 82   1.63432849   0.8530329
#> 83  -0.76641302   0.3069658
#> 84   0.45594932   0.5850015
#> 85  -0.89572075   0.2775538
#> 86   0.83506330   0.6712339
#> 87   0.55392110   0.6072860
#> 88  -1.41807838   0.1587395
#> 89  -1.12658380   0.2250422
#> 90   0.55918097   0.6084824
#> 91  -0.80889915   0.2973020
#> 92   1.93645104   0.9217530
#> 93  -0.21700943   0.4319319
#> 94   0.28521860   0.5461675
#> 95   0.24699511   0.5374733
#> 96  -0.21844781   0.4316047
#> 97  -0.45794314   0.3771297
#> 98   0.32764899   0.5558186
#> 99   0.49571925   0.5940475
#> 100  1.68528659   0.8646237

iris %>% mutate(Petal.Length2 = maxmin(Petal.Length))
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species Petal.Length2
#> 1            5.1         3.5          1.4         0.2     setosa    0.06779661
#> 2            4.9         3.0          1.4         0.2     setosa    0.06779661
#> 3            4.7         3.2          1.3         0.2     setosa    0.05084746
#> 4            4.6         3.1          1.5         0.2     setosa    0.08474576
#> 5            5.0         3.6          1.4         0.2     setosa    0.06779661
#> 6            5.4         3.9          1.7         0.4     setosa    0.11864407
#> 7            4.6         3.4          1.4         0.3     setosa    0.06779661
#> 8            5.0         3.4          1.5         0.2     setosa    0.08474576
#> 9            4.4         2.9          1.4         0.2     setosa    0.06779661
#> 10           4.9         3.1          1.5         0.1     setosa    0.08474576
#> 11           5.4         3.7          1.5         0.2     setosa    0.08474576
#> 12           4.8         3.4          1.6         0.2     setosa    0.10169492
#> 13           4.8         3.0          1.4         0.1     setosa    0.06779661
#> 14           4.3         3.0          1.1         0.1     setosa    0.01694915
#> 15           5.8         4.0          1.2         0.2     setosa    0.03389831
#> 16           5.7         4.4          1.5         0.4     setosa    0.08474576
#> 17           5.4         3.9          1.3         0.4     setosa    0.05084746
#> 18           5.1         3.5          1.4         0.3     setosa    0.06779661
#> 19           5.7         3.8          1.7         0.3     setosa    0.11864407
#> 20           5.1         3.8          1.5         0.3     setosa    0.08474576
#> 21           5.4         3.4          1.7         0.2     setosa    0.11864407
#> 22           5.1         3.7          1.5         0.4     setosa    0.08474576
#> 23           4.6         3.6          1.0         0.2     setosa    0.00000000
#> 24           5.1         3.3          1.7         0.5     setosa    0.11864407
#> 25           4.8         3.4          1.9         0.2     setosa    0.15254237
#> 26           5.0         3.0          1.6         0.2     setosa    0.10169492
#> 27           5.0         3.4          1.6         0.4     setosa    0.10169492
#> 28           5.2         3.5          1.5         0.2     setosa    0.08474576
#> 29           5.2         3.4          1.4         0.2     setosa    0.06779661
#> 30           4.7         3.2          1.6         0.2     setosa    0.10169492
#> 31           4.8         3.1          1.6         0.2     setosa    0.10169492
#> 32           5.4         3.4          1.5         0.4     setosa    0.08474576
#> 33           5.2         4.1          1.5         0.1     setosa    0.08474576
#> 34           5.5         4.2          1.4         0.2     setosa    0.06779661
#> 35           4.9         3.1          1.5         0.2     setosa    0.08474576
#> 36           5.0         3.2          1.2         0.2     setosa    0.03389831
#> 37           5.5         3.5          1.3         0.2     setosa    0.05084746
#> 38           4.9         3.6          1.4         0.1     setosa    0.06779661
#> 39           4.4         3.0          1.3         0.2     setosa    0.05084746
#> 40           5.1         3.4          1.5         0.2     setosa    0.08474576
#> 41           5.0         3.5          1.3         0.3     setosa    0.05084746
#> 42           4.5         2.3          1.3         0.3     setosa    0.05084746
#> 43           4.4         3.2          1.3         0.2     setosa    0.05084746
#> 44           5.0         3.5          1.6         0.6     setosa    0.10169492
#> 45           5.1         3.8          1.9         0.4     setosa    0.15254237
#> 46           4.8         3.0          1.4         0.3     setosa    0.06779661
#> 47           5.1         3.8          1.6         0.2     setosa    0.10169492
#> 48           4.6         3.2          1.4         0.2     setosa    0.06779661
#> 49           5.3         3.7          1.5         0.2     setosa    0.08474576
#> 50           5.0         3.3          1.4         0.2     setosa    0.06779661
#> 51           7.0         3.2          4.7         1.4 versicolor    0.62711864
#> 52           6.4         3.2          4.5         1.5 versicolor    0.59322034
#> 53           6.9         3.1          4.9         1.5 versicolor    0.66101695
#> 54           5.5         2.3          4.0         1.3 versicolor    0.50847458
#> 55           6.5         2.8          4.6         1.5 versicolor    0.61016949
#> 56           5.7         2.8          4.5         1.3 versicolor    0.59322034
#> 57           6.3         3.3          4.7         1.6 versicolor    0.62711864
#> 58           4.9         2.4          3.3         1.0 versicolor    0.38983051
#> 59           6.6         2.9          4.6         1.3 versicolor    0.61016949
#> 60           5.2         2.7          3.9         1.4 versicolor    0.49152542
#> 61           5.0         2.0          3.5         1.0 versicolor    0.42372881
#> 62           5.9         3.0          4.2         1.5 versicolor    0.54237288
#> 63           6.0         2.2          4.0         1.0 versicolor    0.50847458
#> 64           6.1         2.9          4.7         1.4 versicolor    0.62711864
#> 65           5.6         2.9          3.6         1.3 versicolor    0.44067797
#> 66           6.7         3.1          4.4         1.4 versicolor    0.57627119
#> 67           5.6         3.0          4.5         1.5 versicolor    0.59322034
#> 68           5.8         2.7          4.1         1.0 versicolor    0.52542373
#> 69           6.2         2.2          4.5         1.5 versicolor    0.59322034
#> 70           5.6         2.5          3.9         1.1 versicolor    0.49152542
#> 71           5.9         3.2          4.8         1.8 versicolor    0.64406780
#> 72           6.1         2.8          4.0         1.3 versicolor    0.50847458
#> 73           6.3         2.5          4.9         1.5 versicolor    0.66101695
#> 74           6.1         2.8          4.7         1.2 versicolor    0.62711864
#> 75           6.4         2.9          4.3         1.3 versicolor    0.55932203
#> 76           6.6         3.0          4.4         1.4 versicolor    0.57627119
#> 77           6.8         2.8          4.8         1.4 versicolor    0.64406780
#> 78           6.7         3.0          5.0         1.7 versicolor    0.67796610
#> 79           6.0         2.9          4.5         1.5 versicolor    0.59322034
#> 80           5.7         2.6          3.5         1.0 versicolor    0.42372881
#> 81           5.5         2.4          3.8         1.1 versicolor    0.47457627
#> 82           5.5         2.4          3.7         1.0 versicolor    0.45762712
#> 83           5.8         2.7          3.9         1.2 versicolor    0.49152542
#> 84           6.0         2.7          5.1         1.6 versicolor    0.69491525
#> 85           5.4         3.0          4.5         1.5 versicolor    0.59322034
#> 86           6.0         3.4          4.5         1.6 versicolor    0.59322034
#> 87           6.7         3.1          4.7         1.5 versicolor    0.62711864
#> 88           6.3         2.3          4.4         1.3 versicolor    0.57627119
#> 89           5.6         3.0          4.1         1.3 versicolor    0.52542373
#> 90           5.5         2.5          4.0         1.3 versicolor    0.50847458
#> 91           5.5         2.6          4.4         1.2 versicolor    0.57627119
#> 92           6.1         3.0          4.6         1.4 versicolor    0.61016949
#> 93           5.8         2.6          4.0         1.2 versicolor    0.50847458
#> 94           5.0         2.3          3.3         1.0 versicolor    0.38983051
#> 95           5.6         2.7          4.2         1.3 versicolor    0.54237288
#> 96           5.7         3.0          4.2         1.2 versicolor    0.54237288
#> 97           5.7         2.9          4.2         1.3 versicolor    0.54237288
#> 98           6.2         2.9          4.3         1.3 versicolor    0.55932203
#> 99           5.1         2.5          3.0         1.1 versicolor    0.33898305
#> 100          5.7         2.8          4.1         1.3 versicolor    0.52542373
#> 101          6.3         3.3          6.0         2.5  virginica    0.84745763
#> 102          5.8         2.7          5.1         1.9  virginica    0.69491525
#> 103          7.1         3.0          5.9         2.1  virginica    0.83050847
#> 104          6.3         2.9          5.6         1.8  virginica    0.77966102
#> 105          6.5         3.0          5.8         2.2  virginica    0.81355932
#> 106          7.6         3.0          6.6         2.1  virginica    0.94915254
#> 107          4.9         2.5          4.5         1.7  virginica    0.59322034
#> 108          7.3         2.9          6.3         1.8  virginica    0.89830508
#> 109          6.7         2.5          5.8         1.8  virginica    0.81355932
#> 110          7.2         3.6          6.1         2.5  virginica    0.86440678
#> 111          6.5         3.2          5.1         2.0  virginica    0.69491525
#> 112          6.4         2.7          5.3         1.9  virginica    0.72881356
#> 113          6.8         3.0          5.5         2.1  virginica    0.76271186
#> 114          5.7         2.5          5.0         2.0  virginica    0.67796610
#> 115          5.8         2.8          5.1         2.4  virginica    0.69491525
#> 116          6.4         3.2          5.3         2.3  virginica    0.72881356
#> 117          6.5         3.0          5.5         1.8  virginica    0.76271186
#> 118          7.7         3.8          6.7         2.2  virginica    0.96610169
#> 119          7.7         2.6          6.9         2.3  virginica    1.00000000
#> 120          6.0         2.2          5.0         1.5  virginica    0.67796610
#> 121          6.9         3.2          5.7         2.3  virginica    0.79661017
#> 122          5.6         2.8          4.9         2.0  virginica    0.66101695
#> 123          7.7         2.8          6.7         2.0  virginica    0.96610169
#> 124          6.3         2.7          4.9         1.8  virginica    0.66101695
#> 125          6.7         3.3          5.7         2.1  virginica    0.79661017
#> 126          7.2         3.2          6.0         1.8  virginica    0.84745763
#> 127          6.2         2.8          4.8         1.8  virginica    0.64406780
#> 128          6.1         3.0          4.9         1.8  virginica    0.66101695
#> 129          6.4         2.8          5.6         2.1  virginica    0.77966102
#> 130          7.2         3.0          5.8         1.6  virginica    0.81355932
#> 131          7.4         2.8          6.1         1.9  virginica    0.86440678
#> 132          7.9         3.8          6.4         2.0  virginica    0.91525424
#> 133          6.4         2.8          5.6         2.2  virginica    0.77966102
#> 134          6.3         2.8          5.1         1.5  virginica    0.69491525
#> 135          6.1         2.6          5.6         1.4  virginica    0.77966102
#> 136          7.7         3.0          6.1         2.3  virginica    0.86440678
#> 137          6.3         3.4          5.6         2.4  virginica    0.77966102
#> 138          6.4         3.1          5.5         1.8  virginica    0.76271186
#> 139          6.0         3.0          4.8         1.8  virginica    0.64406780
#> 140          6.9         3.1          5.4         2.1  virginica    0.74576271
#> 141          6.7         3.1          5.6         2.4  virginica    0.77966102
#> 142          6.9         3.1          5.1         2.3  virginica    0.69491525
#> 143          5.8         2.7          5.1         1.9  virginica    0.69491525
#> 144          6.8         3.2          5.9         2.3  virginica    0.83050847
#> 145          6.7         3.3          5.7         2.5  virginica    0.79661017
#> 146          6.7         3.0          5.2         2.3  virginica    0.71186441
#> 147          6.3         2.5          5.0         1.9  virginica    0.67796610
#> 148          6.5         3.0          5.2         2.0  virginica    0.71186441
#> 149          6.2         3.4          5.4         2.3  virginica    0.74576271
#> 150          5.9         3.0          5.1         1.8  virginica    0.69491525

maxmin(iris$Petal.Length)
#>   [1] 0.06779661 0.06779661 0.05084746 0.08474576 0.06779661 0.11864407
#>   [7] 0.06779661 0.08474576 0.06779661 0.08474576 0.08474576 0.10169492
#>  [13] 0.06779661 0.01694915 0.03389831 0.08474576 0.05084746 0.06779661
#>  [19] 0.11864407 0.08474576 0.11864407 0.08474576 0.00000000 0.11864407
#>  [25] 0.15254237 0.10169492 0.10169492 0.08474576 0.06779661 0.10169492
#>  [31] 0.10169492 0.08474576 0.08474576 0.06779661 0.08474576 0.03389831
#>  [37] 0.05084746 0.06779661 0.05084746 0.08474576 0.05084746 0.05084746
#>  [43] 0.05084746 0.10169492 0.15254237 0.06779661 0.10169492 0.06779661
#>  [49] 0.08474576 0.06779661 0.62711864 0.59322034 0.66101695 0.50847458
#>  [55] 0.61016949 0.59322034 0.62711864 0.38983051 0.61016949 0.49152542
#>  [61] 0.42372881 0.54237288 0.50847458 0.62711864 0.44067797 0.57627119
#>  [67] 0.59322034 0.52542373 0.59322034 0.49152542 0.64406780 0.50847458
#>  [73] 0.66101695 0.62711864 0.55932203 0.57627119 0.64406780 0.67796610
#>  [79] 0.59322034 0.42372881 0.47457627 0.45762712 0.49152542 0.69491525
#>  [85] 0.59322034 0.59322034 0.62711864 0.57627119 0.52542373 0.50847458
#>  [91] 0.57627119 0.61016949 0.50847458 0.38983051 0.54237288 0.54237288
#>  [97] 0.54237288 0.55932203 0.33898305 0.52542373 0.84745763 0.69491525
#> [103] 0.83050847 0.77966102 0.81355932 0.94915254 0.59322034 0.89830508
#> [109] 0.81355932 0.86440678 0.69491525 0.72881356 0.76271186 0.67796610
#> [115] 0.69491525 0.72881356 0.76271186 0.96610169 1.00000000 0.67796610
#> [121] 0.79661017 0.66101695 0.96610169 0.66101695 0.79661017 0.84745763
#> [127] 0.64406780 0.66101695 0.77966102 0.81355932 0.86440678 0.91525424
#> [133] 0.77966102 0.69491525 0.77966102 0.86440678 0.77966102 0.76271186
#> [139] 0.64406780 0.74576271 0.77966102 0.69491525 0.69491525 0.83050847
#> [145] 0.79661017 0.71186441 0.67796610 0.71186441 0.74576271 0.69491525