Set variable labels

A pipe-workflow optimised method to set variable labels.

set_varl(x, variable_label)

Arguments

x
Variable to assign variable labels to

variable_label
String vector to be assigned as the variable label

See also

Examples

library(tibble)
library(dplyr)
library(magrittr)
df <-
tibble(RESPID=1:1000,
       Q1=sample(c(0,1,2),1000,replace=TRUE),
       Q2=sample(c(0,1),1000,replace=TRUE))
df %>%
  mutate_at("Q1",funs(set_varl(.,"Which of the following groups do you fall into?"))) %>%
  .$Q1

#>    [1] 1 1 2 1 0 0 0 2 1 0 0 2 0 1 0 1 0 0 2 2 0 1 2 1 0 0 1 2 1 2 2 2 0 2 0 1 2
#>   [38] 1 0 0 1 1 2 1 1 1 1 2 1 2 0 0 1 0 2 1 2 0 0 0 1 1 2 2 0 1 2 2 0 2 1 0 2 0
#>   [75] 2 1 2 2 1 2 0 0 0 1 2 2 0 2 0 0 1 1 0 1 1 0 1 1 0 0 1 1 2 2 2 2 2 0 1 0 2
#>  [112] 2 1 0 1 0 1 1 0 0 1 0 1 0 2 2 2 0 1 2 1 1 1 1 0 2 1 2 1 1 0 0 2 1 1 2 0 1
#>  [149] 1 1 0 0 1 2 0 0 2 0 1 1 0 0 2 1 0 1 2 1 0 0 2 2 2 1 0 2 0 0 1 2 2 1 2 0 2
#>  [186] 0 2 2 1 1 0 1 0 1 0 0 1 1 0 0 0 2 0 2 0 1 2 0 0 0 0 2 1 0 0 0 1 2 2 1 0 1
#>  [223] 0 2 1 2 0 1 1 1 2 1 0 1 1 1 0 0 2 1 1 1 0 0 0 1 0 0 1 1 1 0 1 1 2 1 1 1 1
#>  [260] 0 2 0 2 1 2 2 0 2 1 0 2 2 1 2 1 1 1 2 0 0 2 0 0 0 1 2 0 0 0 2 2 0 0 2 0 0
#>  [297] 1 1 2 1 1 0 1 2 0 2 0 2 1 1 1 2 2 1 0 2 1 1 1 2 2 2 0 0 1 2 0 1 1 1 0 0 1
#>  [334] 1 1 1 0 1 2 2 2 2 2 1 2 0 1 1 0 0 2 2 2 2 2 1 2 2 0 2 2 0 1 1 1 2 2 0 2 0
#>  [371] 2 0 2 0 2 0 1 0 0 2 1 1 0 0 0 1 2 1 1 1 2 0 0 2 1 1 0 0 1 0 2 1 0 0 1 0 0
#>  [408] 2 1 1 1 0 0 2 1 0 0 0 1 1 2 0 1 0 2 0 2 2 1 1 1 0 0 0 1 0 2 2 1 1 2 2 0 2
#>  [445] 2 0 1 0 2 1 1 0 0 1 2 0 2 1 1 0 2 0 1 0 2 1 2 0 0 1 0 1 1 1 0 0 2 0 1 1 2
#>  [482] 0 1 2 0 0 0 2 1 0 1 0 0 1 1 1 1 0 1 0 1 0 1 2 2 2 1 0 1 0 0 0 1 1 2 2 0 1
#>  [519] 1 1 1 2 0 0 1 1 1 2 2 1 0 0 1 1 1 0 0 2 1 1 0 1 0 2 0 0 1 2 2 1 1 1 1 1 2
#>  [556] 1 0 0 0 1 2 0 1 1 2 1 2 2 2 1 0 1 1 0 1 1 0 2 1 0 1 0 0 2 0 1 2 2 1 0 0 2
#>  [593] 2 2 1 0 0 2 0 1 0 1 2 0 0 1 1 2 2 0 1 2 1 0 2 2 1 1 2 0 2 1 2 2 0 2 2 0 0
#>  [630] 1 0 2 1 0 2 1 0 1 0 1 2 0 2 1 1 2 0 0 0 0 0 0 0 0 0 1 2 0 0 1 0 1 0 0 1 0
#>  [667] 0 0 1 0 0 2 1 1 0 2 1 1 1 2 1 0 0 1 1 1 1 2 1 1 0 1 0 1 1 1 2 0 2 1 0 2 2
#>  [704] 2 0 1 1 0 2 1 2 2 2 2 0 0 0 1 2 0 1 2 2 1 1 1 2 1 2 0 2 0 2 0 0 0 1 0 2 1
#>  [741] 0 0 1 1 1 1 0 0 2 1 0 2 0 2 0 0 1 0 0 1 2 2 2 1 0 1 1 2 0 0 0 0 2 2 2 1 2
#>  [778] 2 1 2 0 0 0 2 1 2 1 1 0 1 1 0 0 1 0 2 0 1 2 2 1 1 0 0 1 0 1 2 1 1 1 2 2 2
#>  [815] 1 2 0 2 1 2 0 2 1 0 0 2 1 0 2 2 2 1 2 0 2 1 0 1 0 2 2 2 2 1 0 0 0 2 2 1 0
#>  [852] 2 0 0 2 2 0 2 2 1 0 2 0 0 0 1 2 2 0 2 2 1 1 1 0 1 1 0 2 2 0 1 1 0 2 2 1 2
#>  [889] 1 0 1 2 1 0 2 1 2 0 1 0 1 1 0 1 1 0 0 2 0 2 0 1 2 1 2 0 2 1 1 2 1 2 1 1 1
#>  [926] 1 1 2 2 1 2 1 0 2 2 2 0 0 0 1 0 0 2 2 1 1 0 0 1 1 2 0 1 2 1 2 1 0 1 0 1 0
#>  [963] 1 0 1 2 1 1 0 0 2 2 1 2 0 0 2 1 2 2 0 2 1 1 2 1 1 2 2 1 0 2 0 2 2 0 2 1 0
#> [1000] 0
#> attr(,"label")
#> [1] "Which of the following groups do you fall into?"