Max-Min Scaling Function

This function allows you to scale vectors or an entire data frame using the max-min scaling method A numeric vector is always returned.

maxmin(x)

Arguments

x

Pass a vector or the required columns of a data frame through this argument.

Examples

library(magrittr)
library(dplyr)

rand <- rnorm(100, mean = 0, sd = 1)
data.frame(original = rand, transformed = maxmin(rand))
#>        original transformed
#> 1    0.13186850  0.50199381
#> 2    0.36848578  0.54171800
#> 3    3.08913722  0.99847100
#> 4    0.79989370  0.61414435
#> 5    0.87265178  0.62635924
#> 6   -0.18563921  0.44868945
#> 7   -1.06190138  0.30157932
#> 8    0.62824889  0.58532798
#> 9   -0.84462038  0.33805726
#> 10  -0.51282090  0.39376098
#> 11   0.72448675  0.60148475
#> 12   0.59587561  0.57989304
#> 13  -0.64144566  0.37216698
#> 14  -1.26176078  0.26802619
#> 15  -0.34334397  0.42221339
#> 16   1.29384892  0.69707137
#> 17   0.93920094  0.63753176
#> 18   1.70497288  0.76609238
#> 19  -1.32360444  0.25764365
#> 20  -2.41167603  0.07497417
#> 21  -1.06434058  0.30116982
#> 22  -0.45754883  0.40304026
#> 23   0.53650402  0.56992552
#> 24  -0.49813436  0.39622661
#> 25  -2.31776101  0.09074097
#> 26   0.05621956  0.48929359
#> 27   1.85817134  0.79181190
#> 28   1.31818787  0.70115748
#> 29   2.08677802  0.83019124
#> 30  -0.81619284  0.34282978
#> 31   1.43087663  0.72007609
#> 32  -0.33686079  0.42330181
#> 33   0.21536898  0.51601218
#> 34   0.26822036  0.52488507
#> 35   0.64878402  0.58877550
#> 36  -0.15465875  0.45389056
#> 37   2.17551364  0.84508850
#> 38  -1.74995591  0.18606619
#> 39   0.20911761  0.51496268
#> 40   1.90180610  0.79913747
#> 41  -0.58349570  0.38189583
#> 42  -0.44775523  0.40468444
#> 43  -0.76582066  0.35128645
#> 44  -1.21241247  0.27631097
#> 45  -0.38243007  0.41565147
#> 46   0.41839186  0.55009641
#> 47  -0.19585997  0.44697355
#> 48   0.24071700  0.52026770
#> 49   0.28943891  0.52844731
#> 50   0.35894236  0.54011581
#> 51   0.15265411  0.50548338
#> 52  -0.57047094  0.38408248
#> 53   0.29168764  0.52882484
#> 54   0.74975346  0.60572662
#> 55  -1.97729132  0.14790028
#> 56  -0.40837410  0.41129589
#> 57   0.06341669  0.49050187
#> 58  -0.08581264  0.46544870
#> 59  -2.44315009  0.06969019
#> 60   0.01131333  0.48175457
#> 61   0.16932648  0.50828240
#> 62  -0.39444995  0.41363353
#> 63  -2.85826006  0.00000000
#> 64   0.15289397  0.50552365
#> 65  -0.05368974  0.47084161
#> 66  -1.12780736  0.29051478
#> 67   0.72366458  0.60134672
#> 68   0.26136506  0.52373417
#> 69   1.79378998  0.78100332
#> 70  -0.90337358  0.32819355
#> 71   1.44545810  0.72252408
#> 72  -0.12753917  0.45844350
#> 73  -0.02665713  0.47537995
#> 74   0.68763341  0.59529768
#> 75   3.09824474  1.00000000
#> 76  -0.11270228  0.46093437
#> 77  -0.13748939  0.45677302
#> 78  -0.90767496  0.32747142
#> 79   0.10764166  0.49792652
#> 80   1.83861420  0.78852858
#> 81   1.08077705  0.66130008
#> 82   2.21141511  0.85111577
#> 83  -1.25947714  0.26840957
#> 84   0.67043404  0.59241018
#> 85  -0.73638526  0.35622817
#> 86   0.09401361  0.49563860
#> 87   0.29314384  0.52906931
#> 88   0.67692060  0.59349917
#> 89  -0.63519890  0.37321571
#> 90  -1.02725804  0.30739537
#> 91   0.49440528  0.56285783
#> 92   0.48739362  0.56168068
#> 93  -0.49988302  0.39593304
#> 94   1.62015275  0.75185246
#> 95  -1.35788242  0.25188893
#> 96  -0.82473222  0.34139616
#> 97  -0.18565371  0.44868701
#> 98   0.60212493  0.58094220
#> 99   0.74957648  0.60569691
#> 100  0.23688930  0.51962509

iris %>% mutate(Petal.Length2 = maxmin(Petal.Length))
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species Petal.Length2
#> 1            5.1         3.5          1.4         0.2     setosa    0.06779661
#> 2            4.9         3.0          1.4         0.2     setosa    0.06779661
#> 3            4.7         3.2          1.3         0.2     setosa    0.05084746
#> 4            4.6         3.1          1.5         0.2     setosa    0.08474576
#> 5            5.0         3.6          1.4         0.2     setosa    0.06779661
#> 6            5.4         3.9          1.7         0.4     setosa    0.11864407
#> 7            4.6         3.4          1.4         0.3     setosa    0.06779661
#> 8            5.0         3.4          1.5         0.2     setosa    0.08474576
#> 9            4.4         2.9          1.4         0.2     setosa    0.06779661
#> 10           4.9         3.1          1.5         0.1     setosa    0.08474576
#> 11           5.4         3.7          1.5         0.2     setosa    0.08474576
#> 12           4.8         3.4          1.6         0.2     setosa    0.10169492
#> 13           4.8         3.0          1.4         0.1     setosa    0.06779661
#> 14           4.3         3.0          1.1         0.1     setosa    0.01694915
#> 15           5.8         4.0          1.2         0.2     setosa    0.03389831
#> 16           5.7         4.4          1.5         0.4     setosa    0.08474576
#> 17           5.4         3.9          1.3         0.4     setosa    0.05084746
#> 18           5.1         3.5          1.4         0.3     setosa    0.06779661
#> 19           5.7         3.8          1.7         0.3     setosa    0.11864407
#> 20           5.1         3.8          1.5         0.3     setosa    0.08474576
#> 21           5.4         3.4          1.7         0.2     setosa    0.11864407
#> 22           5.1         3.7          1.5         0.4     setosa    0.08474576
#> 23           4.6         3.6          1.0         0.2     setosa    0.00000000
#> 24           5.1         3.3          1.7         0.5     setosa    0.11864407
#> 25           4.8         3.4          1.9         0.2     setosa    0.15254237
#> 26           5.0         3.0          1.6         0.2     setosa    0.10169492
#> 27           5.0         3.4          1.6         0.4     setosa    0.10169492
#> 28           5.2         3.5          1.5         0.2     setosa    0.08474576
#> 29           5.2         3.4          1.4         0.2     setosa    0.06779661
#> 30           4.7         3.2          1.6         0.2     setosa    0.10169492
#> 31           4.8         3.1          1.6         0.2     setosa    0.10169492
#> 32           5.4         3.4          1.5         0.4     setosa    0.08474576
#> 33           5.2         4.1          1.5         0.1     setosa    0.08474576
#> 34           5.5         4.2          1.4         0.2     setosa    0.06779661
#> 35           4.9         3.1          1.5         0.2     setosa    0.08474576
#> 36           5.0         3.2          1.2         0.2     setosa    0.03389831
#> 37           5.5         3.5          1.3         0.2     setosa    0.05084746
#> 38           4.9         3.6          1.4         0.1     setosa    0.06779661
#> 39           4.4         3.0          1.3         0.2     setosa    0.05084746
#> 40           5.1         3.4          1.5         0.2     setosa    0.08474576
#> 41           5.0         3.5          1.3         0.3     setosa    0.05084746
#> 42           4.5         2.3          1.3         0.3     setosa    0.05084746
#> 43           4.4         3.2          1.3         0.2     setosa    0.05084746
#> 44           5.0         3.5          1.6         0.6     setosa    0.10169492
#> 45           5.1         3.8          1.9         0.4     setosa    0.15254237
#> 46           4.8         3.0          1.4         0.3     setosa    0.06779661
#> 47           5.1         3.8          1.6         0.2     setosa    0.10169492
#> 48           4.6         3.2          1.4         0.2     setosa    0.06779661
#> 49           5.3         3.7          1.5         0.2     setosa    0.08474576
#> 50           5.0         3.3          1.4         0.2     setosa    0.06779661
#> 51           7.0         3.2          4.7         1.4 versicolor    0.62711864
#> 52           6.4         3.2          4.5         1.5 versicolor    0.59322034
#> 53           6.9         3.1          4.9         1.5 versicolor    0.66101695
#> 54           5.5         2.3          4.0         1.3 versicolor    0.50847458
#> 55           6.5         2.8          4.6         1.5 versicolor    0.61016949
#> 56           5.7         2.8          4.5         1.3 versicolor    0.59322034
#> 57           6.3         3.3          4.7         1.6 versicolor    0.62711864
#> 58           4.9         2.4          3.3         1.0 versicolor    0.38983051
#> 59           6.6         2.9          4.6         1.3 versicolor    0.61016949
#> 60           5.2         2.7          3.9         1.4 versicolor    0.49152542
#> 61           5.0         2.0          3.5         1.0 versicolor    0.42372881
#> 62           5.9         3.0          4.2         1.5 versicolor    0.54237288
#> 63           6.0         2.2          4.0         1.0 versicolor    0.50847458
#> 64           6.1         2.9          4.7         1.4 versicolor    0.62711864
#> 65           5.6         2.9          3.6         1.3 versicolor    0.44067797
#> 66           6.7         3.1          4.4         1.4 versicolor    0.57627119
#> 67           5.6         3.0          4.5         1.5 versicolor    0.59322034
#> 68           5.8         2.7          4.1         1.0 versicolor    0.52542373
#> 69           6.2         2.2          4.5         1.5 versicolor    0.59322034
#> 70           5.6         2.5          3.9         1.1 versicolor    0.49152542
#> 71           5.9         3.2          4.8         1.8 versicolor    0.64406780
#> 72           6.1         2.8          4.0         1.3 versicolor    0.50847458
#> 73           6.3         2.5          4.9         1.5 versicolor    0.66101695
#> 74           6.1         2.8          4.7         1.2 versicolor    0.62711864
#> 75           6.4         2.9          4.3         1.3 versicolor    0.55932203
#> 76           6.6         3.0          4.4         1.4 versicolor    0.57627119
#> 77           6.8         2.8          4.8         1.4 versicolor    0.64406780
#> 78           6.7         3.0          5.0         1.7 versicolor    0.67796610
#> 79           6.0         2.9          4.5         1.5 versicolor    0.59322034
#> 80           5.7         2.6          3.5         1.0 versicolor    0.42372881
#> 81           5.5         2.4          3.8         1.1 versicolor    0.47457627
#> 82           5.5         2.4          3.7         1.0 versicolor    0.45762712
#> 83           5.8         2.7          3.9         1.2 versicolor    0.49152542
#> 84           6.0         2.7          5.1         1.6 versicolor    0.69491525
#> 85           5.4         3.0          4.5         1.5 versicolor    0.59322034
#> 86           6.0         3.4          4.5         1.6 versicolor    0.59322034
#> 87           6.7         3.1          4.7         1.5 versicolor    0.62711864
#> 88           6.3         2.3          4.4         1.3 versicolor    0.57627119
#> 89           5.6         3.0          4.1         1.3 versicolor    0.52542373
#> 90           5.5         2.5          4.0         1.3 versicolor    0.50847458
#> 91           5.5         2.6          4.4         1.2 versicolor    0.57627119
#> 92           6.1         3.0          4.6         1.4 versicolor    0.61016949
#> 93           5.8         2.6          4.0         1.2 versicolor    0.50847458
#> 94           5.0         2.3          3.3         1.0 versicolor    0.38983051
#> 95           5.6         2.7          4.2         1.3 versicolor    0.54237288
#> 96           5.7         3.0          4.2         1.2 versicolor    0.54237288
#> 97           5.7         2.9          4.2         1.3 versicolor    0.54237288
#> 98           6.2         2.9          4.3         1.3 versicolor    0.55932203
#> 99           5.1         2.5          3.0         1.1 versicolor    0.33898305
#> 100          5.7         2.8          4.1         1.3 versicolor    0.52542373
#> 101          6.3         3.3          6.0         2.5  virginica    0.84745763
#> 102          5.8         2.7          5.1         1.9  virginica    0.69491525
#> 103          7.1         3.0          5.9         2.1  virginica    0.83050847
#> 104          6.3         2.9          5.6         1.8  virginica    0.77966102
#> 105          6.5         3.0          5.8         2.2  virginica    0.81355932
#> 106          7.6         3.0          6.6         2.1  virginica    0.94915254
#> 107          4.9         2.5          4.5         1.7  virginica    0.59322034
#> 108          7.3         2.9          6.3         1.8  virginica    0.89830508
#> 109          6.7         2.5          5.8         1.8  virginica    0.81355932
#> 110          7.2         3.6          6.1         2.5  virginica    0.86440678
#> 111          6.5         3.2          5.1         2.0  virginica    0.69491525
#> 112          6.4         2.7          5.3         1.9  virginica    0.72881356
#> 113          6.8         3.0          5.5         2.1  virginica    0.76271186
#> 114          5.7         2.5          5.0         2.0  virginica    0.67796610
#> 115          5.8         2.8          5.1         2.4  virginica    0.69491525
#> 116          6.4         3.2          5.3         2.3  virginica    0.72881356
#> 117          6.5         3.0          5.5         1.8  virginica    0.76271186
#> 118          7.7         3.8          6.7         2.2  virginica    0.96610169
#> 119          7.7         2.6          6.9         2.3  virginica    1.00000000
#> 120          6.0         2.2          5.0         1.5  virginica    0.67796610
#> 121          6.9         3.2          5.7         2.3  virginica    0.79661017
#> 122          5.6         2.8          4.9         2.0  virginica    0.66101695
#> 123          7.7         2.8          6.7         2.0  virginica    0.96610169
#> 124          6.3         2.7          4.9         1.8  virginica    0.66101695
#> 125          6.7         3.3          5.7         2.1  virginica    0.79661017
#> 126          7.2         3.2          6.0         1.8  virginica    0.84745763
#> 127          6.2         2.8          4.8         1.8  virginica    0.64406780
#> 128          6.1         3.0          4.9         1.8  virginica    0.66101695
#> 129          6.4         2.8          5.6         2.1  virginica    0.77966102
#> 130          7.2         3.0          5.8         1.6  virginica    0.81355932
#> 131          7.4         2.8          6.1         1.9  virginica    0.86440678
#> 132          7.9         3.8          6.4         2.0  virginica    0.91525424
#> 133          6.4         2.8          5.6         2.2  virginica    0.77966102
#> 134          6.3         2.8          5.1         1.5  virginica    0.69491525
#> 135          6.1         2.6          5.6         1.4  virginica    0.77966102
#> 136          7.7         3.0          6.1         2.3  virginica    0.86440678
#> 137          6.3         3.4          5.6         2.4  virginica    0.77966102
#> 138          6.4         3.1          5.5         1.8  virginica    0.76271186
#> 139          6.0         3.0          4.8         1.8  virginica    0.64406780
#> 140          6.9         3.1          5.4         2.1  virginica    0.74576271
#> 141          6.7         3.1          5.6         2.4  virginica    0.77966102
#> 142          6.9         3.1          5.1         2.3  virginica    0.69491525
#> 143          5.8         2.7          5.1         1.9  virginica    0.69491525
#> 144          6.8         3.2          5.9         2.3  virginica    0.83050847
#> 145          6.7         3.3          5.7         2.5  virginica    0.79661017
#> 146          6.7         3.0          5.2         2.3  virginica    0.71186441
#> 147          6.3         2.5          5.0         1.9  virginica    0.67796610
#> 148          6.5         3.0          5.2         2.0  virginica    0.71186441
#> 149          6.2         3.4          5.4         2.3  virginica    0.74576271
#> 150          5.9         3.0          5.1         1.8  virginica    0.69491525

maxmin(iris$Petal.Length)
#>   [1] 0.06779661 0.06779661 0.05084746 0.08474576 0.06779661 0.11864407
#>   [7] 0.06779661 0.08474576 0.06779661 0.08474576 0.08474576 0.10169492
#>  [13] 0.06779661 0.01694915 0.03389831 0.08474576 0.05084746 0.06779661
#>  [19] 0.11864407 0.08474576 0.11864407 0.08474576 0.00000000 0.11864407
#>  [25] 0.15254237 0.10169492 0.10169492 0.08474576 0.06779661 0.10169492
#>  [31] 0.10169492 0.08474576 0.08474576 0.06779661 0.08474576 0.03389831
#>  [37] 0.05084746 0.06779661 0.05084746 0.08474576 0.05084746 0.05084746
#>  [43] 0.05084746 0.10169492 0.15254237 0.06779661 0.10169492 0.06779661
#>  [49] 0.08474576 0.06779661 0.62711864 0.59322034 0.66101695 0.50847458
#>  [55] 0.61016949 0.59322034 0.62711864 0.38983051 0.61016949 0.49152542
#>  [61] 0.42372881 0.54237288 0.50847458 0.62711864 0.44067797 0.57627119
#>  [67] 0.59322034 0.52542373 0.59322034 0.49152542 0.64406780 0.50847458
#>  [73] 0.66101695 0.62711864 0.55932203 0.57627119 0.64406780 0.67796610
#>  [79] 0.59322034 0.42372881 0.47457627 0.45762712 0.49152542 0.69491525
#>  [85] 0.59322034 0.59322034 0.62711864 0.57627119 0.52542373 0.50847458
#>  [91] 0.57627119 0.61016949 0.50847458 0.38983051 0.54237288 0.54237288
#>  [97] 0.54237288 0.55932203 0.33898305 0.52542373 0.84745763 0.69491525
#> [103] 0.83050847 0.77966102 0.81355932 0.94915254 0.59322034 0.89830508
#> [109] 0.81355932 0.86440678 0.69491525 0.72881356 0.76271186 0.67796610
#> [115] 0.69491525 0.72881356 0.76271186 0.96610169 1.00000000 0.67796610
#> [121] 0.79661017 0.66101695 0.96610169 0.66101695 0.79661017 0.84745763
#> [127] 0.64406780 0.66101695 0.77966102 0.81355932 0.86440678 0.91525424
#> [133] 0.77966102 0.69491525 0.77966102 0.86440678 0.77966102 0.76271186
#> [139] 0.64406780 0.74576271 0.77966102 0.69491525 0.69491525 0.83050847
#> [145] 0.79661017 0.71186441 0.67796610 0.71186441 0.74576271 0.69491525