Max-Min Scaling Function

This function allows you to scale vectors or an entire data frame using the max-min scaling method A numeric vector is always returned.

maxmin(x)

Arguments

x

Pass a vector or the required columns of a data frame through this argument.

Examples

library(magrittr)
library(dplyr)

rand <- rnorm(100, mean = 0, sd = 1)
data.frame(original = rand, transformed = maxmin(rand))
#>         original transformed
#> 1   -0.453741023  0.34249077
#> 2    0.459955091  0.51258744
#> 3    0.889665329  0.59258371
#> 4    0.620377305  0.54245217
#> 5   -0.029048469  0.42155293
#> 6   -0.895908063  0.26017550
#> 7    0.407126692  0.50275273
#> 8   -0.263646142  0.37787946
#> 9    0.212889451  0.46659289
#> 10  -0.785973921  0.28064120
#> 11  -0.300630735  0.37099429
#> 12  -0.580987400  0.31880217
#> 13   0.386645653  0.49893992
#> 14   0.452644735  0.51122652
#> 15  -1.638046152  0.12201664
#> 16   1.340890479  0.67658527
#> 17   0.398876956  0.50121694
#> 18  -1.916238755  0.07022739
#> 19  -0.722113436  0.29252968
#> 20   3.078154264  1.00000000
#> 21   0.530271381  0.52567775
#> 22  -1.224283354  0.19904408
#> 23  -1.302338504  0.18451308
#> 24   0.492313353  0.51861136
#> 25  -0.628417770  0.30997237
#> 26   0.330877566  0.48855795
#> 27  -0.153974356  0.39829632
#> 28   1.075359277  0.62715311
#> 29   0.386880385  0.49898361
#> 30  -0.216967603  0.38656929
#> 31  -2.013138752  0.05218817
#> 32  -0.270493003  0.37660482
#> 33  -0.007096288  0.42563962
#> 34   1.856015288  0.77248259
#> 35   0.227838813  0.46937591
#> 36  -1.512046020  0.14547324
#> 37   0.757151824  0.56791456
#> 38   0.373523635  0.49649708
#> 39  -0.151734698  0.39871326
#> 40  -0.849339728  0.26884482
#> 41   0.450831691  0.51088900
#> 42   0.317296150  0.48602958
#> 43   1.300484465  0.66906315
#> 44   0.295394409  0.48195228
#> 45   0.454723168  0.51161345
#> 46  -0.552252712  0.32415151
#> 47  -2.293474198  0.00000000
#> 48   0.031555769  0.43283522
#> 49   0.583554509  0.53559711
#> 50   0.071944698  0.44035415
#> 51  -1.121238231  0.21822730
#> 52  -1.986054897  0.05723019
#> 53   0.668097636  0.55133594
#> 54  -0.009182621  0.42525122
#> 55  -0.665314572  0.30310354
#> 56  -1.946261519  0.06463825
#> 57  -0.541246572  0.32620045
#> 58   0.071721563  0.44031261
#> 59  -0.402092337  0.35210586
#> 60   1.448190743  0.69656064
#> 61  -1.494596900  0.14872162
#> 62  -0.314521941  0.36840825
#> 63   0.746913925  0.56600864
#> 64  -0.720407775  0.29284721
#> 65  -0.637829620  0.30822023
#> 66  -0.734910893  0.29014726
#> 67   1.508712202  0.70782751
#> 68   0.677363571  0.55306092
#> 69   0.347596422  0.49167038
#> 70   0.080526765  0.44195182
#> 71   0.480363143  0.51638667
#> 72  -0.894346620  0.26046619
#> 73   1.738764562  0.75065481
#> 74   1.249029991  0.65948422
#> 75   0.375153576  0.49680051
#> 76   1.304718912  0.66985145
#> 77   0.749967361  0.56657708
#> 78   0.104175965  0.44635443
#> 79   0.534656433  0.52649409
#> 80   0.874240474  0.58971217
#> 81   1.283248902  0.66585452
#> 82   1.207457651  0.65174497
#> 83  -1.327295568  0.17986699
#> 84  -0.236695199  0.38289673
#> 85  -0.286983375  0.37353492
#> 86  -0.215110162  0.38691508
#> 87  -1.464232177  0.15437442
#> 88  -0.938901829  0.25217164
#> 89  -0.888817905  0.26149543
#> 90  -0.247566134  0.38087297
#> 91  -0.764616610  0.28461715
#> 92   0.286182859  0.48023743
#> 93  -1.479563543  0.15152028
#> 94   1.983607683  0.79623561
#> 95  -1.093905245  0.22331570
#> 96   0.767526837  0.56984601
#> 97  -0.319405129  0.36749918
#> 98   1.234574609  0.65679316
#> 99  -0.007913482  0.42548749
#> 100 -1.311919358  0.18272947

iris %>% mutate(Petal.Length2 = maxmin(Petal.Length))
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species Petal.Length2
#> 1            5.1         3.5          1.4         0.2     setosa    0.06779661
#> 2            4.9         3.0          1.4         0.2     setosa    0.06779661
#> 3            4.7         3.2          1.3         0.2     setosa    0.05084746
#> 4            4.6         3.1          1.5         0.2     setosa    0.08474576
#> 5            5.0         3.6          1.4         0.2     setosa    0.06779661
#> 6            5.4         3.9          1.7         0.4     setosa    0.11864407
#> 7            4.6         3.4          1.4         0.3     setosa    0.06779661
#> 8            5.0         3.4          1.5         0.2     setosa    0.08474576
#> 9            4.4         2.9          1.4         0.2     setosa    0.06779661
#> 10           4.9         3.1          1.5         0.1     setosa    0.08474576
#> 11           5.4         3.7          1.5         0.2     setosa    0.08474576
#> 12           4.8         3.4          1.6         0.2     setosa    0.10169492
#> 13           4.8         3.0          1.4         0.1     setosa    0.06779661
#> 14           4.3         3.0          1.1         0.1     setosa    0.01694915
#> 15           5.8         4.0          1.2         0.2     setosa    0.03389831
#> 16           5.7         4.4          1.5         0.4     setosa    0.08474576
#> 17           5.4         3.9          1.3         0.4     setosa    0.05084746
#> 18           5.1         3.5          1.4         0.3     setosa    0.06779661
#> 19           5.7         3.8          1.7         0.3     setosa    0.11864407
#> 20           5.1         3.8          1.5         0.3     setosa    0.08474576
#> 21           5.4         3.4          1.7         0.2     setosa    0.11864407
#> 22           5.1         3.7          1.5         0.4     setosa    0.08474576
#> 23           4.6         3.6          1.0         0.2     setosa    0.00000000
#> 24           5.1         3.3          1.7         0.5     setosa    0.11864407
#> 25           4.8         3.4          1.9         0.2     setosa    0.15254237
#> 26           5.0         3.0          1.6         0.2     setosa    0.10169492
#> 27           5.0         3.4          1.6         0.4     setosa    0.10169492
#> 28           5.2         3.5          1.5         0.2     setosa    0.08474576
#> 29           5.2         3.4          1.4         0.2     setosa    0.06779661
#> 30           4.7         3.2          1.6         0.2     setosa    0.10169492
#> 31           4.8         3.1          1.6         0.2     setosa    0.10169492
#> 32           5.4         3.4          1.5         0.4     setosa    0.08474576
#> 33           5.2         4.1          1.5         0.1     setosa    0.08474576
#> 34           5.5         4.2          1.4         0.2     setosa    0.06779661
#> 35           4.9         3.1          1.5         0.2     setosa    0.08474576
#> 36           5.0         3.2          1.2         0.2     setosa    0.03389831
#> 37           5.5         3.5          1.3         0.2     setosa    0.05084746
#> 38           4.9         3.6          1.4         0.1     setosa    0.06779661
#> 39           4.4         3.0          1.3         0.2     setosa    0.05084746
#> 40           5.1         3.4          1.5         0.2     setosa    0.08474576
#> 41           5.0         3.5          1.3         0.3     setosa    0.05084746
#> 42           4.5         2.3          1.3         0.3     setosa    0.05084746
#> 43           4.4         3.2          1.3         0.2     setosa    0.05084746
#> 44           5.0         3.5          1.6         0.6     setosa    0.10169492
#> 45           5.1         3.8          1.9         0.4     setosa    0.15254237
#> 46           4.8         3.0          1.4         0.3     setosa    0.06779661
#> 47           5.1         3.8          1.6         0.2     setosa    0.10169492
#> 48           4.6         3.2          1.4         0.2     setosa    0.06779661
#> 49           5.3         3.7          1.5         0.2     setosa    0.08474576
#> 50           5.0         3.3          1.4         0.2     setosa    0.06779661
#> 51           7.0         3.2          4.7         1.4 versicolor    0.62711864
#> 52           6.4         3.2          4.5         1.5 versicolor    0.59322034
#> 53           6.9         3.1          4.9         1.5 versicolor    0.66101695
#> 54           5.5         2.3          4.0         1.3 versicolor    0.50847458
#> 55           6.5         2.8          4.6         1.5 versicolor    0.61016949
#> 56           5.7         2.8          4.5         1.3 versicolor    0.59322034
#> 57           6.3         3.3          4.7         1.6 versicolor    0.62711864
#> 58           4.9         2.4          3.3         1.0 versicolor    0.38983051
#> 59           6.6         2.9          4.6         1.3 versicolor    0.61016949
#> 60           5.2         2.7          3.9         1.4 versicolor    0.49152542
#> 61           5.0         2.0          3.5         1.0 versicolor    0.42372881
#> 62           5.9         3.0          4.2         1.5 versicolor    0.54237288
#> 63           6.0         2.2          4.0         1.0 versicolor    0.50847458
#> 64           6.1         2.9          4.7         1.4 versicolor    0.62711864
#> 65           5.6         2.9          3.6         1.3 versicolor    0.44067797
#> 66           6.7         3.1          4.4         1.4 versicolor    0.57627119
#> 67           5.6         3.0          4.5         1.5 versicolor    0.59322034
#> 68           5.8         2.7          4.1         1.0 versicolor    0.52542373
#> 69           6.2         2.2          4.5         1.5 versicolor    0.59322034
#> 70           5.6         2.5          3.9         1.1 versicolor    0.49152542
#> 71           5.9         3.2          4.8         1.8 versicolor    0.64406780
#> 72           6.1         2.8          4.0         1.3 versicolor    0.50847458
#> 73           6.3         2.5          4.9         1.5 versicolor    0.66101695
#> 74           6.1         2.8          4.7         1.2 versicolor    0.62711864
#> 75           6.4         2.9          4.3         1.3 versicolor    0.55932203
#> 76           6.6         3.0          4.4         1.4 versicolor    0.57627119
#> 77           6.8         2.8          4.8         1.4 versicolor    0.64406780
#> 78           6.7         3.0          5.0         1.7 versicolor    0.67796610
#> 79           6.0         2.9          4.5         1.5 versicolor    0.59322034
#> 80           5.7         2.6          3.5         1.0 versicolor    0.42372881
#> 81           5.5         2.4          3.8         1.1 versicolor    0.47457627
#> 82           5.5         2.4          3.7         1.0 versicolor    0.45762712
#> 83           5.8         2.7          3.9         1.2 versicolor    0.49152542
#> 84           6.0         2.7          5.1         1.6 versicolor    0.69491525
#> 85           5.4         3.0          4.5         1.5 versicolor    0.59322034
#> 86           6.0         3.4          4.5         1.6 versicolor    0.59322034
#> 87           6.7         3.1          4.7         1.5 versicolor    0.62711864
#> 88           6.3         2.3          4.4         1.3 versicolor    0.57627119
#> 89           5.6         3.0          4.1         1.3 versicolor    0.52542373
#> 90           5.5         2.5          4.0         1.3 versicolor    0.50847458
#> 91           5.5         2.6          4.4         1.2 versicolor    0.57627119
#> 92           6.1         3.0          4.6         1.4 versicolor    0.61016949
#> 93           5.8         2.6          4.0         1.2 versicolor    0.50847458
#> 94           5.0         2.3          3.3         1.0 versicolor    0.38983051
#> 95           5.6         2.7          4.2         1.3 versicolor    0.54237288
#> 96           5.7         3.0          4.2         1.2 versicolor    0.54237288
#> 97           5.7         2.9          4.2         1.3 versicolor    0.54237288
#> 98           6.2         2.9          4.3         1.3 versicolor    0.55932203
#> 99           5.1         2.5          3.0         1.1 versicolor    0.33898305
#> 100          5.7         2.8          4.1         1.3 versicolor    0.52542373
#> 101          6.3         3.3          6.0         2.5  virginica    0.84745763
#> 102          5.8         2.7          5.1         1.9  virginica    0.69491525
#> 103          7.1         3.0          5.9         2.1  virginica    0.83050847
#> 104          6.3         2.9          5.6         1.8  virginica    0.77966102
#> 105          6.5         3.0          5.8         2.2  virginica    0.81355932
#> 106          7.6         3.0          6.6         2.1  virginica    0.94915254
#> 107          4.9         2.5          4.5         1.7  virginica    0.59322034
#> 108          7.3         2.9          6.3         1.8  virginica    0.89830508
#> 109          6.7         2.5          5.8         1.8  virginica    0.81355932
#> 110          7.2         3.6          6.1         2.5  virginica    0.86440678
#> 111          6.5         3.2          5.1         2.0  virginica    0.69491525
#> 112          6.4         2.7          5.3         1.9  virginica    0.72881356
#> 113          6.8         3.0          5.5         2.1  virginica    0.76271186
#> 114          5.7         2.5          5.0         2.0  virginica    0.67796610
#> 115          5.8         2.8          5.1         2.4  virginica    0.69491525
#> 116          6.4         3.2          5.3         2.3  virginica    0.72881356
#> 117          6.5         3.0          5.5         1.8  virginica    0.76271186
#> 118          7.7         3.8          6.7         2.2  virginica    0.96610169
#> 119          7.7         2.6          6.9         2.3  virginica    1.00000000
#> 120          6.0         2.2          5.0         1.5  virginica    0.67796610
#> 121          6.9         3.2          5.7         2.3  virginica    0.79661017
#> 122          5.6         2.8          4.9         2.0  virginica    0.66101695
#> 123          7.7         2.8          6.7         2.0  virginica    0.96610169
#> 124          6.3         2.7          4.9         1.8  virginica    0.66101695
#> 125          6.7         3.3          5.7         2.1  virginica    0.79661017
#> 126          7.2         3.2          6.0         1.8  virginica    0.84745763
#> 127          6.2         2.8          4.8         1.8  virginica    0.64406780
#> 128          6.1         3.0          4.9         1.8  virginica    0.66101695
#> 129          6.4         2.8          5.6         2.1  virginica    0.77966102
#> 130          7.2         3.0          5.8         1.6  virginica    0.81355932
#> 131          7.4         2.8          6.1         1.9  virginica    0.86440678
#> 132          7.9         3.8          6.4         2.0  virginica    0.91525424
#> 133          6.4         2.8          5.6         2.2  virginica    0.77966102
#> 134          6.3         2.8          5.1         1.5  virginica    0.69491525
#> 135          6.1         2.6          5.6         1.4  virginica    0.77966102
#> 136          7.7         3.0          6.1         2.3  virginica    0.86440678
#> 137          6.3         3.4          5.6         2.4  virginica    0.77966102
#> 138          6.4         3.1          5.5         1.8  virginica    0.76271186
#> 139          6.0         3.0          4.8         1.8  virginica    0.64406780
#> 140          6.9         3.1          5.4         2.1  virginica    0.74576271
#> 141          6.7         3.1          5.6         2.4  virginica    0.77966102
#> 142          6.9         3.1          5.1         2.3  virginica    0.69491525
#> 143          5.8         2.7          5.1         1.9  virginica    0.69491525
#> 144          6.8         3.2          5.9         2.3  virginica    0.83050847
#> 145          6.7         3.3          5.7         2.5  virginica    0.79661017
#> 146          6.7         3.0          5.2         2.3  virginica    0.71186441
#> 147          6.3         2.5          5.0         1.9  virginica    0.67796610
#> 148          6.5         3.0          5.2         2.0  virginica    0.71186441
#> 149          6.2         3.4          5.4         2.3  virginica    0.74576271
#> 150          5.9         3.0          5.1         1.8  virginica    0.69491525

maxmin(iris$Petal.Length)
#>   [1] 0.06779661 0.06779661 0.05084746 0.08474576 0.06779661 0.11864407
#>   [7] 0.06779661 0.08474576 0.06779661 0.08474576 0.08474576 0.10169492
#>  [13] 0.06779661 0.01694915 0.03389831 0.08474576 0.05084746 0.06779661
#>  [19] 0.11864407 0.08474576 0.11864407 0.08474576 0.00000000 0.11864407
#>  [25] 0.15254237 0.10169492 0.10169492 0.08474576 0.06779661 0.10169492
#>  [31] 0.10169492 0.08474576 0.08474576 0.06779661 0.08474576 0.03389831
#>  [37] 0.05084746 0.06779661 0.05084746 0.08474576 0.05084746 0.05084746
#>  [43] 0.05084746 0.10169492 0.15254237 0.06779661 0.10169492 0.06779661
#>  [49] 0.08474576 0.06779661 0.62711864 0.59322034 0.66101695 0.50847458
#>  [55] 0.61016949 0.59322034 0.62711864 0.38983051 0.61016949 0.49152542
#>  [61] 0.42372881 0.54237288 0.50847458 0.62711864 0.44067797 0.57627119
#>  [67] 0.59322034 0.52542373 0.59322034 0.49152542 0.64406780 0.50847458
#>  [73] 0.66101695 0.62711864 0.55932203 0.57627119 0.64406780 0.67796610
#>  [79] 0.59322034 0.42372881 0.47457627 0.45762712 0.49152542 0.69491525
#>  [85] 0.59322034 0.59322034 0.62711864 0.57627119 0.52542373 0.50847458
#>  [91] 0.57627119 0.61016949 0.50847458 0.38983051 0.54237288 0.54237288
#>  [97] 0.54237288 0.55932203 0.33898305 0.52542373 0.84745763 0.69491525
#> [103] 0.83050847 0.77966102 0.81355932 0.94915254 0.59322034 0.89830508
#> [109] 0.81355932 0.86440678 0.69491525 0.72881356 0.76271186 0.67796610
#> [115] 0.69491525 0.72881356 0.76271186 0.96610169 1.00000000 0.67796610
#> [121] 0.79661017 0.66101695 0.96610169 0.66101695 0.79661017 0.84745763
#> [127] 0.64406780 0.66101695 0.77966102 0.81355932 0.86440678 0.91525424
#> [133] 0.77966102 0.69491525 0.77966102 0.86440678 0.77966102 0.76271186
#> [139] 0.64406780 0.74576271 0.77966102 0.69491525 0.69491525 0.83050847
#> [145] 0.79661017 0.71186441 0.67796610 0.71186441 0.74576271 0.69491525