Max-Min Scaling Function

This function allows you to scale vectors or an entire data frame using the max-min scaling method A numeric vector is always returned.

maxmin(x)

Arguments

x

Pass a vector or the required columns of a data frame through this argument.

Examples

library(magrittr)
library(dplyr)

rand <- rnorm(100, mean = 0, sd = 1)
data.frame(original = rand, transformed = maxmin(rand))
#>        original transformed
#> 1    1.14981796  0.77594710
#> 2   -0.78765332  0.42310901
#> 3    0.26826200  0.61540458
#> 4   -0.20518300  0.52918424
#> 5    0.74279756  0.70182352
#> 6    0.60254205  0.67628122
#> 7    2.38011613  1.00000000
#> 8   -1.31955826  0.32624238
#> 9    0.44171575  0.64699271
#> 10   0.48254579  0.65442838
#> 11  -0.06953347  0.55388774
#> 12  -3.11098966  0.00000000
#> 13  -0.45764019  0.48320859
#> 14   0.94524439  0.73869166
#> 15   0.49736982  0.65712802
#> 16   1.00899808  0.75030201
#> 17   1.45581230  0.83167255
#> 18   0.16440494  0.59649089
#> 19   1.15712644  0.77727807
#> 20  -1.66889663  0.26262343
#> 21   1.06073733  0.75972439
#> 22  -0.95837288  0.39201882
#> 23   0.16214491  0.59607931
#> 24  -0.56929115  0.46287553
#> 25  -0.58084816  0.46077085
#> 26   0.42246842  0.64348753
#> 27   0.61784179  0.67906749
#> 28   1.21528694  0.78786983
#> 29  -0.24716022  0.52153966
#> 30   0.32504800  0.62574603
#> 31   0.12181530  0.58873478
#> 32  -0.84467626  0.41272441
#> 33  -0.27483548  0.51649964
#> 34  -0.51018477  0.47363955
#> 35  -0.21746201  0.52694808
#> 36  -0.78985095  0.42270879
#> 37   1.04141961  0.75620639
#> 38  -0.56182863  0.46423455
#> 39   1.71301275  0.87851202
#> 40  -1.99601312  0.20305137
#> 41   1.99990100  0.93075800
#> 42   1.82143948  0.89825790
#> 43  -0.04069326  0.55913991
#> 44   0.10361558  0.58542038
#> 45  -0.05611024  0.55633228
#> 46   0.65931815  0.68662087
#> 47  -2.70473999  0.07398322
#> 48   2.02559271  0.93543679
#> 49   0.32928153  0.62651701
#> 50   0.22049993  0.60670650
#> 51   0.71712234  0.69714774
#> 52  -0.44852785  0.48486806
#> 53  -0.93385633  0.39648359
#> 54  -0.51832345  0.47215740
#> 55  -0.20462133  0.52928653
#> 56  -0.31780813  0.50867378
#> 57   1.69439265  0.87512106
#> 58   0.70124243  0.69425581
#> 59   1.91679281  0.91562295
#> 60   0.42656997  0.64423447
#> 61   1.27621744  0.79896605
#> 62  -1.56728288  0.28112858
#> 63  -0.17839215  0.53406320
#> 64   1.49148765  0.83816948
#> 65  -0.32366781  0.50760666
#> 66  -1.06796656  0.37206042
#> 67   1.95495285  0.92257237
#> 68   1.11640855  0.76986282
#> 69  -0.93389982  0.39647567
#> 70   0.95855161  0.74111507
#> 71  -1.16993174  0.35349126
#> 72   0.15300981  0.59441570
#> 73   1.62575299  0.86262091
#> 74   0.28591075  0.61861864
#> 75   0.43740442  0.64620756
#> 76   0.33465417  0.62749544
#> 77   1.18127035  0.78167498
#> 78   1.12056451  0.77061968
#> 79   0.11604166  0.58768333
#> 80  -1.19385897  0.34913381
#> 81  -1.21538317  0.34521398
#> 82  -0.82542523  0.41623027
#> 83  -0.75722635  0.42865015
#> 84   1.48913248  0.83774058
#> 85   0.46455201  0.65115148
#> 86  -1.07346210  0.37105961
#> 87  -0.57718823  0.46143737
#> 88   0.97502478  0.74411505
#> 89  -1.01613502  0.38149960
#> 90  -0.49954551  0.47557710
#> 91   1.96017430  0.92352327
#> 92   0.22447541  0.60743049
#> 93   0.74576980  0.70236481
#> 94  -0.61232238  0.45503900
#> 95   1.40661441  0.82271299
#> 96   1.23247927  0.79100077
#> 97  -1.13110282  0.36056250
#> 98   0.08498987  0.58202840
#> 99  -1.79562504  0.23954458
#> 100  0.58996016  0.67398990

iris %>% mutate(Petal.Length2 = maxmin(Petal.Length))
#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species Petal.Length2
#> 1            5.1         3.5          1.4         0.2     setosa    0.06779661
#> 2            4.9         3.0          1.4         0.2     setosa    0.06779661
#> 3            4.7         3.2          1.3         0.2     setosa    0.05084746
#> 4            4.6         3.1          1.5         0.2     setosa    0.08474576
#> 5            5.0         3.6          1.4         0.2     setosa    0.06779661
#> 6            5.4         3.9          1.7         0.4     setosa    0.11864407
#> 7            4.6         3.4          1.4         0.3     setosa    0.06779661
#> 8            5.0         3.4          1.5         0.2     setosa    0.08474576
#> 9            4.4         2.9          1.4         0.2     setosa    0.06779661
#> 10           4.9         3.1          1.5         0.1     setosa    0.08474576
#> 11           5.4         3.7          1.5         0.2     setosa    0.08474576
#> 12           4.8         3.4          1.6         0.2     setosa    0.10169492
#> 13           4.8         3.0          1.4         0.1     setosa    0.06779661
#> 14           4.3         3.0          1.1         0.1     setosa    0.01694915
#> 15           5.8         4.0          1.2         0.2     setosa    0.03389831
#> 16           5.7         4.4          1.5         0.4     setosa    0.08474576
#> 17           5.4         3.9          1.3         0.4     setosa    0.05084746
#> 18           5.1         3.5          1.4         0.3     setosa    0.06779661
#> 19           5.7         3.8          1.7         0.3     setosa    0.11864407
#> 20           5.1         3.8          1.5         0.3     setosa    0.08474576
#> 21           5.4         3.4          1.7         0.2     setosa    0.11864407
#> 22           5.1         3.7          1.5         0.4     setosa    0.08474576
#> 23           4.6         3.6          1.0         0.2     setosa    0.00000000
#> 24           5.1         3.3          1.7         0.5     setosa    0.11864407
#> 25           4.8         3.4          1.9         0.2     setosa    0.15254237
#> 26           5.0         3.0          1.6         0.2     setosa    0.10169492
#> 27           5.0         3.4          1.6         0.4     setosa    0.10169492
#> 28           5.2         3.5          1.5         0.2     setosa    0.08474576
#> 29           5.2         3.4          1.4         0.2     setosa    0.06779661
#> 30           4.7         3.2          1.6         0.2     setosa    0.10169492
#> 31           4.8         3.1          1.6         0.2     setosa    0.10169492
#> 32           5.4         3.4          1.5         0.4     setosa    0.08474576
#> 33           5.2         4.1          1.5         0.1     setosa    0.08474576
#> 34           5.5         4.2          1.4         0.2     setosa    0.06779661
#> 35           4.9         3.1          1.5         0.2     setosa    0.08474576
#> 36           5.0         3.2          1.2         0.2     setosa    0.03389831
#> 37           5.5         3.5          1.3         0.2     setosa    0.05084746
#> 38           4.9         3.6          1.4         0.1     setosa    0.06779661
#> 39           4.4         3.0          1.3         0.2     setosa    0.05084746
#> 40           5.1         3.4          1.5         0.2     setosa    0.08474576
#> 41           5.0         3.5          1.3         0.3     setosa    0.05084746
#> 42           4.5         2.3          1.3         0.3     setosa    0.05084746
#> 43           4.4         3.2          1.3         0.2     setosa    0.05084746
#> 44           5.0         3.5          1.6         0.6     setosa    0.10169492
#> 45           5.1         3.8          1.9         0.4     setosa    0.15254237
#> 46           4.8         3.0          1.4         0.3     setosa    0.06779661
#> 47           5.1         3.8          1.6         0.2     setosa    0.10169492
#> 48           4.6         3.2          1.4         0.2     setosa    0.06779661
#> 49           5.3         3.7          1.5         0.2     setosa    0.08474576
#> 50           5.0         3.3          1.4         0.2     setosa    0.06779661
#> 51           7.0         3.2          4.7         1.4 versicolor    0.62711864
#> 52           6.4         3.2          4.5         1.5 versicolor    0.59322034
#> 53           6.9         3.1          4.9         1.5 versicolor    0.66101695
#> 54           5.5         2.3          4.0         1.3 versicolor    0.50847458
#> 55           6.5         2.8          4.6         1.5 versicolor    0.61016949
#> 56           5.7         2.8          4.5         1.3 versicolor    0.59322034
#> 57           6.3         3.3          4.7         1.6 versicolor    0.62711864
#> 58           4.9         2.4          3.3         1.0 versicolor    0.38983051
#> 59           6.6         2.9          4.6         1.3 versicolor    0.61016949
#> 60           5.2         2.7          3.9         1.4 versicolor    0.49152542
#> 61           5.0         2.0          3.5         1.0 versicolor    0.42372881
#> 62           5.9         3.0          4.2         1.5 versicolor    0.54237288
#> 63           6.0         2.2          4.0         1.0 versicolor    0.50847458
#> 64           6.1         2.9          4.7         1.4 versicolor    0.62711864
#> 65           5.6         2.9          3.6         1.3 versicolor    0.44067797
#> 66           6.7         3.1          4.4         1.4 versicolor    0.57627119
#> 67           5.6         3.0          4.5         1.5 versicolor    0.59322034
#> 68           5.8         2.7          4.1         1.0 versicolor    0.52542373
#> 69           6.2         2.2          4.5         1.5 versicolor    0.59322034
#> 70           5.6         2.5          3.9         1.1 versicolor    0.49152542
#> 71           5.9         3.2          4.8         1.8 versicolor    0.64406780
#> 72           6.1         2.8          4.0         1.3 versicolor    0.50847458
#> 73           6.3         2.5          4.9         1.5 versicolor    0.66101695
#> 74           6.1         2.8          4.7         1.2 versicolor    0.62711864
#> 75           6.4         2.9          4.3         1.3 versicolor    0.55932203
#> 76           6.6         3.0          4.4         1.4 versicolor    0.57627119
#> 77           6.8         2.8          4.8         1.4 versicolor    0.64406780
#> 78           6.7         3.0          5.0         1.7 versicolor    0.67796610
#> 79           6.0         2.9          4.5         1.5 versicolor    0.59322034
#> 80           5.7         2.6          3.5         1.0 versicolor    0.42372881
#> 81           5.5         2.4          3.8         1.1 versicolor    0.47457627
#> 82           5.5         2.4          3.7         1.0 versicolor    0.45762712
#> 83           5.8         2.7          3.9         1.2 versicolor    0.49152542
#> 84           6.0         2.7          5.1         1.6 versicolor    0.69491525
#> 85           5.4         3.0          4.5         1.5 versicolor    0.59322034
#> 86           6.0         3.4          4.5         1.6 versicolor    0.59322034
#> 87           6.7         3.1          4.7         1.5 versicolor    0.62711864
#> 88           6.3         2.3          4.4         1.3 versicolor    0.57627119
#> 89           5.6         3.0          4.1         1.3 versicolor    0.52542373
#> 90           5.5         2.5          4.0         1.3 versicolor    0.50847458
#> 91           5.5         2.6          4.4         1.2 versicolor    0.57627119
#> 92           6.1         3.0          4.6         1.4 versicolor    0.61016949
#> 93           5.8         2.6          4.0         1.2 versicolor    0.50847458
#> 94           5.0         2.3          3.3         1.0 versicolor    0.38983051
#> 95           5.6         2.7          4.2         1.3 versicolor    0.54237288
#> 96           5.7         3.0          4.2         1.2 versicolor    0.54237288
#> 97           5.7         2.9          4.2         1.3 versicolor    0.54237288
#> 98           6.2         2.9          4.3         1.3 versicolor    0.55932203
#> 99           5.1         2.5          3.0         1.1 versicolor    0.33898305
#> 100          5.7         2.8          4.1         1.3 versicolor    0.52542373
#> 101          6.3         3.3          6.0         2.5  virginica    0.84745763
#> 102          5.8         2.7          5.1         1.9  virginica    0.69491525
#> 103          7.1         3.0          5.9         2.1  virginica    0.83050847
#> 104          6.3         2.9          5.6         1.8  virginica    0.77966102
#> 105          6.5         3.0          5.8         2.2  virginica    0.81355932
#> 106          7.6         3.0          6.6         2.1  virginica    0.94915254
#> 107          4.9         2.5          4.5         1.7  virginica    0.59322034
#> 108          7.3         2.9          6.3         1.8  virginica    0.89830508
#> 109          6.7         2.5          5.8         1.8  virginica    0.81355932
#> 110          7.2         3.6          6.1         2.5  virginica    0.86440678
#> 111          6.5         3.2          5.1         2.0  virginica    0.69491525
#> 112          6.4         2.7          5.3         1.9  virginica    0.72881356
#> 113          6.8         3.0          5.5         2.1  virginica    0.76271186
#> 114          5.7         2.5          5.0         2.0  virginica    0.67796610
#> 115          5.8         2.8          5.1         2.4  virginica    0.69491525
#> 116          6.4         3.2          5.3         2.3  virginica    0.72881356
#> 117          6.5         3.0          5.5         1.8  virginica    0.76271186
#> 118          7.7         3.8          6.7         2.2  virginica    0.96610169
#> 119          7.7         2.6          6.9         2.3  virginica    1.00000000
#> 120          6.0         2.2          5.0         1.5  virginica    0.67796610
#> 121          6.9         3.2          5.7         2.3  virginica    0.79661017
#> 122          5.6         2.8          4.9         2.0  virginica    0.66101695
#> 123          7.7         2.8          6.7         2.0  virginica    0.96610169
#> 124          6.3         2.7          4.9         1.8  virginica    0.66101695
#> 125          6.7         3.3          5.7         2.1  virginica    0.79661017
#> 126          7.2         3.2          6.0         1.8  virginica    0.84745763
#> 127          6.2         2.8          4.8         1.8  virginica    0.64406780
#> 128          6.1         3.0          4.9         1.8  virginica    0.66101695
#> 129          6.4         2.8          5.6         2.1  virginica    0.77966102
#> 130          7.2         3.0          5.8         1.6  virginica    0.81355932
#> 131          7.4         2.8          6.1         1.9  virginica    0.86440678
#> 132          7.9         3.8          6.4         2.0  virginica    0.91525424
#> 133          6.4         2.8          5.6         2.2  virginica    0.77966102
#> 134          6.3         2.8          5.1         1.5  virginica    0.69491525
#> 135          6.1         2.6          5.6         1.4  virginica    0.77966102
#> 136          7.7         3.0          6.1         2.3  virginica    0.86440678
#> 137          6.3         3.4          5.6         2.4  virginica    0.77966102
#> 138          6.4         3.1          5.5         1.8  virginica    0.76271186
#> 139          6.0         3.0          4.8         1.8  virginica    0.64406780
#> 140          6.9         3.1          5.4         2.1  virginica    0.74576271
#> 141          6.7         3.1          5.6         2.4  virginica    0.77966102
#> 142          6.9         3.1          5.1         2.3  virginica    0.69491525
#> 143          5.8         2.7          5.1         1.9  virginica    0.69491525
#> 144          6.8         3.2          5.9         2.3  virginica    0.83050847
#> 145          6.7         3.3          5.7         2.5  virginica    0.79661017
#> 146          6.7         3.0          5.2         2.3  virginica    0.71186441
#> 147          6.3         2.5          5.0         1.9  virginica    0.67796610
#> 148          6.5         3.0          5.2         2.0  virginica    0.71186441
#> 149          6.2         3.4          5.4         2.3  virginica    0.74576271
#> 150          5.9         3.0          5.1         1.8  virginica    0.69491525

maxmin(iris$Petal.Length)
#>   [1] 0.06779661 0.06779661 0.05084746 0.08474576 0.06779661 0.11864407
#>   [7] 0.06779661 0.08474576 0.06779661 0.08474576 0.08474576 0.10169492
#>  [13] 0.06779661 0.01694915 0.03389831 0.08474576 0.05084746 0.06779661
#>  [19] 0.11864407 0.08474576 0.11864407 0.08474576 0.00000000 0.11864407
#>  [25] 0.15254237 0.10169492 0.10169492 0.08474576 0.06779661 0.10169492
#>  [31] 0.10169492 0.08474576 0.08474576 0.06779661 0.08474576 0.03389831
#>  [37] 0.05084746 0.06779661 0.05084746 0.08474576 0.05084746 0.05084746
#>  [43] 0.05084746 0.10169492 0.15254237 0.06779661 0.10169492 0.06779661
#>  [49] 0.08474576 0.06779661 0.62711864 0.59322034 0.66101695 0.50847458
#>  [55] 0.61016949 0.59322034 0.62711864 0.38983051 0.61016949 0.49152542
#>  [61] 0.42372881 0.54237288 0.50847458 0.62711864 0.44067797 0.57627119
#>  [67] 0.59322034 0.52542373 0.59322034 0.49152542 0.64406780 0.50847458
#>  [73] 0.66101695 0.62711864 0.55932203 0.57627119 0.64406780 0.67796610
#>  [79] 0.59322034 0.42372881 0.47457627 0.45762712 0.49152542 0.69491525
#>  [85] 0.59322034 0.59322034 0.62711864 0.57627119 0.52542373 0.50847458
#>  [91] 0.57627119 0.61016949 0.50847458 0.38983051 0.54237288 0.54237288
#>  [97] 0.54237288 0.55932203 0.33898305 0.52542373 0.84745763 0.69491525
#> [103] 0.83050847 0.77966102 0.81355932 0.94915254 0.59322034 0.89830508
#> [109] 0.81355932 0.86440678 0.69491525 0.72881356 0.76271186 0.67796610
#> [115] 0.69491525 0.72881356 0.76271186 0.96610169 1.00000000 0.67796610
#> [121] 0.79661017 0.66101695 0.96610169 0.66101695 0.79661017 0.84745763
#> [127] 0.64406780 0.66101695 0.77966102 0.81355932 0.86440678 0.91525424
#> [133] 0.77966102 0.69491525 0.77966102 0.86440678 0.77966102 0.76271186
#> [139] 0.64406780 0.74576271 0.77966102 0.69491525 0.69491525 0.83050847
#> [145] 0.79661017 0.71186441 0.67796610 0.71186441 0.74576271 0.69491525