Max-Min Scaling Function

This function allows you to scale vectors or an entire data frame using the max-min scaling method A numeric vector is always returned.

maxmin(x)

Arguments

x
Pass a vector or the required columns of a data frame through this argument.

Examples

library(magrittr)
library(dplyr)
rand.data <-cbind(sample(1000,234:697),sample(1000,234:697)) %>% as.data.frame()
rand.data %>% mutate_all(~maxmin(.))

#>              V1          V2
#> 1   0.204660588 0.857429719
#> 2   0.095238095 0.673694779
#> 3   0.876393110 0.605421687
#> 4   0.650455927 0.092369478
#> 5   0.047619048 0.157630522
#> 6   0.110435664 0.349397590
#> 7   0.992907801 0.742971888
#> 8   0.166160081 0.033132530
#> 9   0.739614995 0.085341365
#> 10  0.472137791 0.016064257
#> 11  0.976697062 0.529116466
#> 12  0.658561297 0.454819277
#> 13  0.514690983 0.245983936
#> 14  0.063829787 0.446787149
#> 15  0.014184397 0.566265060
#> 16  0.400202634 0.590361446
#> 17  0.089159068 0.403614458
#> 18  0.936170213 0.391566265
#> 19  0.121580547 0.259036145
#> 20  0.539007092 0.004016064
#> 21  0.683890578 0.770080321
#> 22  0.865248227 0.836345382
#> 23  0.309017224 0.943775100
#> 24  0.847011145 0.535140562
#> 25  0.800405268 0.153614458
#> 26  0.238095238 0.793172691
#> 27  0.163120567 0.013052209
#> 28  0.660587639 0.556224900
#> 29  0.566362715 0.840361446
#> 30  0.356636272 0.819277108
#> 31  0.937183384 0.387550201
#> 32  0.803444782 0.912650602
#> 33  0.021276596 0.167670683
#> 34  0.236068896 0.095381526
#> 35  0.938196555 0.778112450
#> 36  0.084093212 0.189759036
#> 37  0.825734549 0.583333333
#> 38  0.286727457 0.687751004
#> 39  0.670719352 0.411646586
#> 40  0.247213779 0.214859438
#> 41  0.388044580 0.030120482
#> 42  0.142857143 0.474899598
#> 43  0.578520770 0.041164659
#> 44  0.288753799 0.710843373
#> 45  0.491388045 0.809236948
#> 46  0.044579534 0.568273092
#> 47  0.233029382 0.461847390
#> 48  0.882472138 0.237951807
#> 49  0.471124620 0.096385542
#> 50  0.403242148 0.486947791
#> 51  0.492401216 0.601405622
#> 52  0.347517730 0.285140562
#> 53  0.149949341 0.876506024
#> 54  0.484295846 0.057228916
#> 55  0.210739615 0.768072289
#> 56  0.325227964 0.287148594
#> 57  0.382978723 0.577309237
#> 58  0.137791287 0.174698795
#> 59  0.688956434 0.205823293
#> 60  0.537993921 0.630522088
#> 61  0.495440729 0.808232932
#> 62  0.048632219 0.074297189
#> 63  0.664640324 0.308232932
#> 64  0.000000000 0.386546185
#> 65  0.339412361 0.817269076
#> 66  0.124620061 0.522088353
#> 67  0.960486322 0.565261044
#> 68  0.666666667 0.311244980
#> 69  0.535967579 0.892570281
#> 70  0.239108409 0.040160643
#> 71  0.080040527 0.517068273
#> 72  0.251266464 0.520080321
#> 73  0.690982776 0.150602410
#> 74  0.139817629 0.624497992
#> 75  0.802431611 0.762048193
#> 76  0.599797366 0.546184739
#> 77  0.929078014 0.686746988
#> 78  0.665653495 1.000000000
#> 79  0.844984802 0.415662651
#> 80  0.943262411 0.485943775
#> 81  0.003039514 0.787148594
#> 82  0.132725431 0.132530120
#> 83  0.508611955 0.602409639
#> 84  0.002026342 0.792168675
#> 85  0.390070922 0.528112450
#> 86  0.526849037 0.000000000
#> 87  0.263424519 0.900602410
#> 88  0.396149949 0.604417671
#> 89  0.028368794 0.252008032
#> 90  0.401215805 0.646586345
#> 91  0.590678825 0.426706827
#> 92  0.873353597 0.539156627
#> 93  0.459979737 0.756024096
#> 94  0.136778116 0.410642570
#> 95  0.004052685 0.714859438
#> 96  0.727456940 0.433734940
#> 97  1.000000000 0.435742972
#> 98  0.374873354 0.219879518
#> 99  0.454913880 0.201807229
#> 100 0.075987842 0.481927711
#> 101 0.195542047 0.261044177
#> 102 0.672745694 0.717871486
#> 103 0.951367781 0.821285141
#> 104 0.475177305 0.483935743
#> 105 0.284701114 0.408634538
#> 106 0.691995947 0.681726908
#> 107 0.024316109 0.069277108
#> 108 0.726443769 0.838353414
#> 109 0.094224924 0.682730924
#> 110 0.880445795 0.942771084
#> 111 0.891590679 0.080321285
#> 112 0.255319149 0.240963855
#> 113 0.597771023 0.645582329
#> 114 0.765957447 0.106425703
#> 115 0.303951368 0.358433735
#> 116 0.952380952 0.176706827
#> 117 0.482269504 0.814257028
#> 118 0.148936170 0.452811245
#> 119 0.190476190 0.182730924
#> 120 0.920972644 0.238955823
#> 121 0.541033435 0.852409639
#> 122 0.392097264 0.190763052
#> 123 0.196555218 0.855421687
#> 124 0.146909828 0.001004016
#> 125 0.556231003 0.247991968
#> 126 0.569402229 0.305220884
#> 127 0.317122594 0.379518072
#> 128 0.707193516 0.866465863
#> 129 0.643363728 0.385542169
#> 130 0.462006079 0.350401606
#> 131 0.524822695 0.661646586
#> 132 0.996960486 0.286144578
#> 133 0.070921986 0.177710843
#> 134 0.316109422 0.082329317
#> 135 0.023302938 0.439759036
#> 136 0.808510638 0.507028112
#> 137 0.168186424 0.854417671
#> 138 0.586626140 0.015060241
#> 139 0.181357649 0.614457831
#> 140 0.826747720 0.019076305
#> 141 0.040526849 0.378514056
#> 142 0.657548126 0.801204819
#> 143 0.543059777 0.398594378
#> 144 0.628166160 0.173694779
#> 145 0.375886525 0.939759036
#> 146 0.735562310 0.327309237
#> 147 0.711246201 0.697791165
#> 148 0.718338399 0.796184739
#> 149 0.038500507 0.734939759
#> 150 0.925025329 0.609437751
#> 151 0.041540020 0.328313253
#> 152 0.054711246 0.524096386
#> 153 0.743667680 0.584337349
#> 154 0.451874367 0.144578313
#> 155 0.750759878 0.791164659
#> 156 0.807497467 0.956827309
#> 157 0.013171226 0.906626506
#> 158 0.839918946 0.330321285
#> 159 0.906788247 0.597389558
#> 160 0.742654509 0.317269076
#> 161 0.831813576 0.553212851
#> 162 0.759878419 0.868473896
#> 163 0.290780142 0.915662651
#> 164 0.012158055 0.138554217
#> 165 0.351570415 0.797188755
#> 166 0.037487335 0.475903614
#> 167 0.922998987 0.551204819
#> 168 0.398176292 0.582329317
#> 169 0.755825735 0.989959839
#> 170 0.357649443 0.592369478
#> 171 0.328267477 0.023092369
#> 172 0.704154002 0.658634538
#> 173 0.057750760 0.065261044
#> 174 0.405268490 0.818273092
#> 175 0.686930091 0.414658635
#> 176 0.420466059 0.621485944
#> 177 0.079027356 0.336345382
#> 178 0.554204661 0.690763052
#> 179 0.010131712 0.352409639
#> 180 0.751773050 0.318273092
#> 181 0.205673759 0.497991968
#> 182 0.289766971 0.467871486
#> 183 0.940222898 0.170682731
#> 184 0.858156028 0.632530120
#> 185 0.383991895 0.543172691
#> 186 0.832826748 0.506024096
#> 187 0.576494428 0.195783133
#> 188 0.456940223 0.477911647
#> 189 0.812563323 0.755020080
#> 190 0.485309017 0.937751004
#> 191 0.551165147 0.480923695
#> 192 0.355623100 0.462851406
#> 193 0.228976697 0.725903614
#> 194 0.032421479 0.786144578
#> 195 0.915906788 0.152610442
#> 196 0.893617021 0.292168675
#> 197 0.262411348 0.929718876
#> 198 0.201621074 0.396586345
#> 199 0.498480243 0.902610442
#> 200 0.463019250 0.031124498
#> 201 0.394123607 0.672690763
#> 202 0.900709220 0.018072289
#> 203 0.950354610 0.945783133
#> 204 0.176291793 0.361445783
#> 205 0.760891591 0.112449799
#> 206 0.784194529 0.549196787
#> 207 0.641337386 0.732931727
#> 208 0.106382979 0.585341365
#> 209 0.694022290 0.365461847
#> 210 0.828774063 0.035140562
#> 211 0.138804458 0.752008032
#> 212 0.623100304 0.491967871
#> 213 0.135764944 0.965863454
#> 214 0.260385005 0.505020080
#> 215 0.676798379 0.114457831
#> 216 0.746707194 0.134538153
#> 217 0.888551165 0.313253012
#> 218 0.067882472 0.745983936
#> 219 0.570415400 0.567269076
#> 220 0.571428571 0.991967871
#> 221 0.315096251 0.133534137
#> 222 0.867274569 0.264056225
#> 223 0.868287741 0.254016064
#> 224 0.065856130 0.272088353
#> 225 0.519756839 0.159638554
#> 226 0.009118541 0.751004016
#> 227 0.321175279 0.616465863
#> 228 0.730496454 0.374497992
#> 229 0.990881459 0.354417671
#> 230 0.259371834 0.100401606
#> 231 0.007092199 0.223895582
#> 232 0.919959473 0.233935743
#> 233 0.237082067 0.935742972
#> 234 0.911854103 0.012048193

sample(1000,234:677) %>% maxmin()

#>   [1] 0.907537688 0.221105528 0.472361809 0.587939698 0.869346734 0.567839196
#>   [7] 0.631155779 0.109547739 0.264321608 0.662311558 0.344723618 0.033165829
#>  [13] 0.953768844 0.493467337 0.320603015 0.325628141 0.464321608 0.194974874
#>  [19] 0.718592965 0.103517588 0.766834171 0.913567839 0.866331658 0.536683417
#>  [25] 0.600000000 0.883417085 0.968844221 0.519597990 0.814070352 0.041206030
#>  [31] 0.203015075 0.010050251 0.651256281 0.443216080 0.736683417 0.445226131
#>  [37] 0.021105528 0.808040201 0.391959799 0.220100503 0.987939698 0.971859296
#>  [43] 0.160804020 0.208040201 0.238190955 0.642211055 0.474371859 0.457286432
#>  [49] 0.093467337 0.723618090 0.712562814 0.819095477 0.890452261 0.686432161
#>  [55] 0.823115578 0.591959799 0.397989950 0.207035176 0.919597990 0.274371859
#>  [61] 0.341708543 0.975879397 0.711557789 0.257286432 0.860301508 0.338693467
#>  [67] 0.698492462 0.660301508 0.607035176 0.154773869 0.645226131 0.473366834
#>  [73] 0.850251256 0.633165829 0.876381910 0.374874372 0.186934673 0.471356784
#>  [79] 0.089447236 0.648241206 0.102512563 0.744723618 0.625125628 0.405025126
#>  [85] 0.191959799 0.124623116 0.613065327 0.349748744 0.449246231 0.910552764
#>  [91] 0.158793970 0.970854271 0.179899497 0.382914573 0.202010050 0.050251256
#>  [97] 0.868341709 0.708542714 0.058291457 0.387939698 0.165829146 0.316582915
#> [103] 0.046231156 0.125628141 0.423115578 0.923618090 0.928643216 0.055276382
#> [109] 0.805025126 0.411055276 1.000000000 0.223115578 0.590954774 0.767839196
#> [115] 0.097487437 0.764824121 0.856281407 0.538693467 0.393969849 0.558793970
#> [121] 0.877386935 0.483417085 0.895477387 0.161809045 0.454271357 0.649246231
#> [127] 0.938693467 0.312562814 0.577889447 0.083417085 0.489447236 0.442211055
#> [133] 0.007035176 0.146733668 0.982914573 0.263316583 0.192964824 0.661306533
#> [139] 0.622110553 0.313567839 0.863316583 0.912562814 0.064321608 0.222110553
#> [145] 0.615075377 0.195979899 0.149748744 0.045226131 0.954773869 0.873366834
#> [151] 0.603015075 0.485427136 0.204020101 0.259296482 0.352763819 0.765829146
#> [157] 0.305527638 0.334673367 0.339698492 0.018090452 0.922613065 0.324623116
#> [163] 0.092462312 0.318592965 0.256281407 0.926633166 0.644221106 0.729648241
#> [169] 0.025125628 0.467336683 0.788944724 0.763819095 0.966834171 0.006030151
#> [175] 0.512562814 0.836180905 0.794974874 0.296482412 0.028140704 0.551758794
#> [181] 0.673366834 0.176884422 0.241206030 0.201005025 0.653266332 0.174874372
#> [187] 0.353768844 0.535678392 0.376884422 0.117587940 0.364824121 0.005025126
#> [193] 0.309547739 0.787939698 0.959798995 0.159798995 0.886432161 0.583919598
#> [199] 0.084422111 0.871356784 0.214070352 0.775879397 0.000000000 0.984924623
#> [205] 0.595979899 0.670351759 0.961809045 0.525628141 0.147738693 0.914572864
#> [211] 0.758793970 0.115577889 0.173869347 0.998994975 0.849246231 0.518592965
#> [217] 0.716582915 0.566834171 0.760804020 0.791959799 0.205025126 0.511557789
#> [223] 0.522613065 0.962814070 0.597989950 0.751758794 0.588944724 0.108542714
#> [229] 0.933668342 0.383919598 0.678391960 0.749748744 0.078391960 0.553768844

iris %>% mutate_at(vars(Petal.Length),~maxmin(.))

#>     Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1            5.1         3.5   0.06779661         0.2     setosa
#> 2            4.9         3.0   0.06779661         0.2     setosa
#> 3            4.7         3.2   0.05084746         0.2     setosa
#> 4            4.6         3.1   0.08474576         0.2     setosa
#> 5            5.0         3.6   0.06779661         0.2     setosa
#> 6            5.4         3.9   0.11864407         0.4     setosa
#> 7            4.6         3.4   0.06779661         0.3     setosa
#> 8            5.0         3.4   0.08474576         0.2     setosa
#> 9            4.4         2.9   0.06779661         0.2     setosa
#> 10           4.9         3.1   0.08474576         0.1     setosa
#> 11           5.4         3.7   0.08474576         0.2     setosa
#> 12           4.8         3.4   0.10169492         0.2     setosa
#> 13           4.8         3.0   0.06779661         0.1     setosa
#> 14           4.3         3.0   0.01694915         0.1     setosa
#> 15           5.8         4.0   0.03389831         0.2     setosa
#> 16           5.7         4.4   0.08474576         0.4     setosa
#> 17           5.4         3.9   0.05084746         0.4     setosa
#> 18           5.1         3.5   0.06779661         0.3     setosa
#> 19           5.7         3.8   0.11864407         0.3     setosa
#> 20           5.1         3.8   0.08474576         0.3     setosa
#> 21           5.4         3.4   0.11864407         0.2     setosa
#> 22           5.1         3.7   0.08474576         0.4     setosa
#> 23           4.6         3.6   0.00000000         0.2     setosa
#> 24           5.1         3.3   0.11864407         0.5     setosa
#> 25           4.8         3.4   0.15254237         0.2     setosa
#> 26           5.0         3.0   0.10169492         0.2     setosa
#> 27           5.0         3.4   0.10169492         0.4     setosa
#> 28           5.2         3.5   0.08474576         0.2     setosa
#> 29           5.2         3.4   0.06779661         0.2     setosa
#> 30           4.7         3.2   0.10169492         0.2     setosa
#> 31           4.8         3.1   0.10169492         0.2     setosa
#> 32           5.4         3.4   0.08474576         0.4     setosa
#> 33           5.2         4.1   0.08474576         0.1     setosa
#> 34           5.5         4.2   0.06779661         0.2     setosa
#> 35           4.9         3.1   0.08474576         0.2     setosa
#> 36           5.0         3.2   0.03389831         0.2     setosa
#> 37           5.5         3.5   0.05084746         0.2     setosa
#> 38           4.9         3.6   0.06779661         0.1     setosa
#> 39           4.4         3.0   0.05084746         0.2     setosa
#> 40           5.1         3.4   0.08474576         0.2     setosa
#> 41           5.0         3.5   0.05084746         0.3     setosa
#> 42           4.5         2.3   0.05084746         0.3     setosa
#> 43           4.4         3.2   0.05084746         0.2     setosa
#> 44           5.0         3.5   0.10169492         0.6     setosa
#> 45           5.1         3.8   0.15254237         0.4     setosa
#> 46           4.8         3.0   0.06779661         0.3     setosa
#> 47           5.1         3.8   0.10169492         0.2     setosa
#> 48           4.6         3.2   0.06779661         0.2     setosa
#> 49           5.3         3.7   0.08474576         0.2     setosa
#> 50           5.0         3.3   0.06779661         0.2     setosa
#> 51           7.0         3.2   0.62711864         1.4 versicolor
#> 52           6.4         3.2   0.59322034         1.5 versicolor
#> 53           6.9         3.1   0.66101695         1.5 versicolor
#> 54           5.5         2.3   0.50847458         1.3 versicolor
#> 55           6.5         2.8   0.61016949         1.5 versicolor
#> 56           5.7         2.8   0.59322034         1.3 versicolor
#> 57           6.3         3.3   0.62711864         1.6 versicolor
#> 58           4.9         2.4   0.38983051         1.0 versicolor
#> 59           6.6         2.9   0.61016949         1.3 versicolor
#> 60           5.2         2.7   0.49152542         1.4 versicolor
#> 61           5.0         2.0   0.42372881         1.0 versicolor
#> 62           5.9         3.0   0.54237288         1.5 versicolor
#> 63           6.0         2.2   0.50847458         1.0 versicolor
#> 64           6.1         2.9   0.62711864         1.4 versicolor
#> 65           5.6         2.9   0.44067797         1.3 versicolor
#> 66           6.7         3.1   0.57627119         1.4 versicolor
#> 67           5.6         3.0   0.59322034         1.5 versicolor
#> 68           5.8         2.7   0.52542373         1.0 versicolor
#> 69           6.2         2.2   0.59322034         1.5 versicolor
#> 70           5.6         2.5   0.49152542         1.1 versicolor
#> 71           5.9         3.2   0.64406780         1.8 versicolor
#> 72           6.1         2.8   0.50847458         1.3 versicolor
#> 73           6.3         2.5   0.66101695         1.5 versicolor
#> 74           6.1         2.8   0.62711864         1.2 versicolor
#> 75           6.4         2.9   0.55932203         1.3 versicolor
#> 76           6.6         3.0   0.57627119         1.4 versicolor
#> 77           6.8         2.8   0.64406780         1.4 versicolor
#> 78           6.7         3.0   0.67796610         1.7 versicolor
#> 79           6.0         2.9   0.59322034         1.5 versicolor
#> 80           5.7         2.6   0.42372881         1.0 versicolor
#> 81           5.5         2.4   0.47457627         1.1 versicolor
#> 82           5.5         2.4   0.45762712         1.0 versicolor
#> 83           5.8         2.7   0.49152542         1.2 versicolor
#> 84           6.0         2.7   0.69491525         1.6 versicolor
#> 85           5.4         3.0   0.59322034         1.5 versicolor
#> 86           6.0         3.4   0.59322034         1.6 versicolor
#> 87           6.7         3.1   0.62711864         1.5 versicolor
#> 88           6.3         2.3   0.57627119         1.3 versicolor
#> 89           5.6         3.0   0.52542373         1.3 versicolor
#> 90           5.5         2.5   0.50847458         1.3 versicolor
#> 91           5.5         2.6   0.57627119         1.2 versicolor
#> 92           6.1         3.0   0.61016949         1.4 versicolor
#> 93           5.8         2.6   0.50847458         1.2 versicolor
#> 94           5.0         2.3   0.38983051         1.0 versicolor
#> 95           5.6         2.7   0.54237288         1.3 versicolor
#> 96           5.7         3.0   0.54237288         1.2 versicolor
#> 97           5.7         2.9   0.54237288         1.3 versicolor
#> 98           6.2         2.9   0.55932203         1.3 versicolor
#> 99           5.1         2.5   0.33898305         1.1 versicolor
#> 100          5.7         2.8   0.52542373         1.3 versicolor
#> 101          6.3         3.3   0.84745763         2.5  virginica
#> 102          5.8         2.7   0.69491525         1.9  virginica
#> 103          7.1         3.0   0.83050847         2.1  virginica
#> 104          6.3         2.9   0.77966102         1.8  virginica
#> 105          6.5         3.0   0.81355932         2.2  virginica
#> 106          7.6         3.0   0.94915254         2.1  virginica
#> 107          4.9         2.5   0.59322034         1.7  virginica
#> 108          7.3         2.9   0.89830508         1.8  virginica
#> 109          6.7         2.5   0.81355932         1.8  virginica
#> 110          7.2         3.6   0.86440678         2.5  virginica
#> 111          6.5         3.2   0.69491525         2.0  virginica
#> 112          6.4         2.7   0.72881356         1.9  virginica
#> 113          6.8         3.0   0.76271186         2.1  virginica
#> 114          5.7         2.5   0.67796610         2.0  virginica
#> 115          5.8         2.8   0.69491525         2.4  virginica
#> 116          6.4         3.2   0.72881356         2.3  virginica
#> 117          6.5         3.0   0.76271186         1.8  virginica
#> 118          7.7         3.8   0.96610169         2.2  virginica
#> 119          7.7         2.6   1.00000000         2.3  virginica
#> 120          6.0         2.2   0.67796610         1.5  virginica
#> 121          6.9         3.2   0.79661017         2.3  virginica
#> 122          5.6         2.8   0.66101695         2.0  virginica
#> 123          7.7         2.8   0.96610169         2.0  virginica
#> 124          6.3         2.7   0.66101695         1.8  virginica
#> 125          6.7         3.3   0.79661017         2.1  virginica
#> 126          7.2         3.2   0.84745763         1.8  virginica
#> 127          6.2         2.8   0.64406780         1.8  virginica
#> 128          6.1         3.0   0.66101695         1.8  virginica
#> 129          6.4         2.8   0.77966102         2.1  virginica
#> 130          7.2         3.0   0.81355932         1.6  virginica
#> 131          7.4         2.8   0.86440678         1.9  virginica
#> 132          7.9         3.8   0.91525424         2.0  virginica
#> 133          6.4         2.8   0.77966102         2.2  virginica
#> 134          6.3         2.8   0.69491525         1.5  virginica
#> 135          6.1         2.6   0.77966102         1.4  virginica
#> 136          7.7         3.0   0.86440678         2.3  virginica
#> 137          6.3         3.4   0.77966102         2.4  virginica
#> 138          6.4         3.1   0.76271186         1.8  virginica
#> 139          6.0         3.0   0.64406780         1.8  virginica
#> 140          6.9         3.1   0.74576271         2.1  virginica
#> 141          6.7         3.1   0.77966102         2.4  virginica
#> 142          6.9         3.1   0.69491525         2.3  virginica
#> 143          5.8         2.7   0.69491525         1.9  virginica
#> 144          6.8         3.2   0.83050847         2.3  virginica
#> 145          6.7         3.3   0.79661017         2.5  virginica
#> 146          6.7         3.0   0.71186441         2.3  virginica
#> 147          6.3         2.5   0.67796610         1.9  virginica
#> 148          6.5         3.0   0.71186441         2.0  virginica
#> 149          6.2         3.4   0.74576271         2.3  virginica
#> 150          5.9         3.0   0.69491525         1.8  virginica

maxmin(iris$Petal.Length)

#>   [1] 0.06779661 0.06779661 0.05084746 0.08474576 0.06779661 0.11864407
#>   [7] 0.06779661 0.08474576 0.06779661 0.08474576 0.08474576 0.10169492
#>  [13] 0.06779661 0.01694915 0.03389831 0.08474576 0.05084746 0.06779661
#>  [19] 0.11864407 0.08474576 0.11864407 0.08474576 0.00000000 0.11864407
#>  [25] 0.15254237 0.10169492 0.10169492 0.08474576 0.06779661 0.10169492
#>  [31] 0.10169492 0.08474576 0.08474576 0.06779661 0.08474576 0.03389831
#>  [37] 0.05084746 0.06779661 0.05084746 0.08474576 0.05084746 0.05084746
#>  [43] 0.05084746 0.10169492 0.15254237 0.06779661 0.10169492 0.06779661
#>  [49] 0.08474576 0.06779661 0.62711864 0.59322034 0.66101695 0.50847458
#>  [55] 0.61016949 0.59322034 0.62711864 0.38983051 0.61016949 0.49152542
#>  [61] 0.42372881 0.54237288 0.50847458 0.62711864 0.44067797 0.57627119
#>  [67] 0.59322034 0.52542373 0.59322034 0.49152542 0.64406780 0.50847458
#>  [73] 0.66101695 0.62711864 0.55932203 0.57627119 0.64406780 0.67796610
#>  [79] 0.59322034 0.42372881 0.47457627 0.45762712 0.49152542 0.69491525
#>  [85] 0.59322034 0.59322034 0.62711864 0.57627119 0.52542373 0.50847458
#>  [91] 0.57627119 0.61016949 0.50847458 0.38983051 0.54237288 0.54237288
#>  [97] 0.54237288 0.55932203 0.33898305 0.52542373 0.84745763 0.69491525
#> [103] 0.83050847 0.77966102 0.81355932 0.94915254 0.59322034 0.89830508
#> [109] 0.81355932 0.86440678 0.69491525 0.72881356 0.76271186 0.67796610
#> [115] 0.69491525 0.72881356 0.76271186 0.96610169 1.00000000 0.67796610
#> [121] 0.79661017 0.66101695 0.96610169 0.66101695 0.79661017 0.84745763
#> [127] 0.64406780 0.66101695 0.77966102 0.81355932 0.86440678 0.91525424
#> [133] 0.77966102 0.69491525 0.77966102 0.86440678 0.77966102 0.76271186
#> [139] 0.64406780 0.74576271 0.77966102 0.69491525 0.69491525 0.83050847
#> [145] 0.79661017 0.71186441 0.67796610 0.71186441 0.74576271 0.69491525